feat(functions): add html check to publish-document and refactor applications-background-jobs
