--- apps/functions/applications-background-jobs/malware-detected/__tests__/test-utils/test-constants.js
@@ -1,6 +1,0 @@
-export const TEST_BLOB_ACCOUNT = 'test-blob-account';
-export const TEST_BLOB_CONTAINER = 'test-blob-container';
-export const TEST_BLOB_GUID = 'a12b3d4a-a123-123a-1a2b-12345abc1abc';
-export const TEST_FUNCTION_DIRECTORY = '/path/to/function';
-export const TEST_FUNCTION_NAME = 'test-function';
-export const TEST_INVOCATION_ID = 'test-invocation-id';
