--- apps/functions/applications-background-jobs/malware-detected/__tests__/test-utils/test-html.js
@@ -1,149 +1,0 @@
-export const INVALID_HTML_STRING = `
-<!DOCTYPE html>
-<html>
-<head>
-    <title>YouTube Iframe Example</title>
-</head>
-<body>
-    <h1>My YouTube Video</h1>
-    <iframe width="560" height="315" src="https://www.youtube.com/embed/hHEvsjxSbU8?si=-eTnuGjPKCN8PRgI" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>
-</body>
-</html>
-`;
-
-/**
- *
- * @param {String} youtubeUrlExample
- * @returns
- */
-export const createTestYoutubeTemplate = (youtubeUrlExample) => `
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>National Infrastructure Planning</title>
-	<style>
-		* {
-			box-sizing: border-box;
-		}
-
-		body {
-			padding: 0;
-			margin: 0;
-			font-family: Arial, Helvetica, sans-serif;
-		}
-
-		header {
-			background-color: #074745;
-			border-bottom: 10px solid #13a19b;
-			padding: 13px 0;
-		}
-
-		main {
-			padding: 30px 0;
-		}
-
-		h1 {
-			color: #0b0c0c;
-			-webkit-font-smoothing: antialiased;
-			-moz-osx-font-smoothing: grayscale;
-			font-weight: 700;
-			font-size: 2rem;
-			line-height: 1.09375;
-			display: block;
-			margin-top: 0;
-			margin-bottom: 30px;
-		}
-
-		@media (min-width: 40.0625em) {
-			h1 {
-				font-size: 3rem;
-				line-height: 1.0416666667;
-			}
-		}
-
-		.container {
-			width: 100%;
-			max-width: 1020px;
-			padding: 0 15px;
-			margin-right: auto;
-			margin-left: auto;
-		}
-
-		@media (min-width: 40.0625em) {
-			.container {
-				padding: 0 30px;
-			}
-		}
-
-		.video-container {
-			position: relative;
-			width: 100%;
-			height: 0;
-			padding-top: 56.25%;
-		}
-
-		.video-container > iframe {
-			position: absolute;
-			top: 0;
-			left: 0;
-			width: 100%;
-			height: 100%;
-		}
-
-		.logo {
-			width: 100%;
-			max-width: 224px;
-		}
-	</style>
-</head>
-
-<body>
-<header>
-	<div class="container">
-		<svg aria-hidden="true" class="logo" fill="none" viewBox="0 0 224 60">
-			<path
-				fill="#fff"
-				d="M30.989 9.737c-.337.652-.63 1.033-.587 1.544.043.51.576.847.826 1.358.228-.587.608-1.087.565-1.489-.044-.402-.696-1.087-.804-1.413ZM39.737 9.737c.337.652.63 1.033.587 1.544-.043.51-.576.847-.826 1.358-.228-.587-.598-1.087-.565-1.489.032-.402.685-1.087.804-1.413ZM25.968 9.357a.587.587 0 0 0 .587-.587.597.597 0 1 0-1.185 0 .588.588 0 0 0 .598.587ZM26.36 10.205a.598.598 0 1 0-1.185-.161.598.598 0 0 0 1.184.16ZM26.642 11.64a.597.597 0 1 0-.642.586.587.587 0 0 0 .587-.587M29.228 6.542a.598.598 0 0 0 .598-.587.598.598 0 0 0-1.185 0 .587.587 0 0 0 .587.587ZM27.761 7.042a.598.598 0 1 0-.587-.598.585.585 0 0 0 .587.598ZM26.609 8.01a.587.587 0 0 0 .609-.566.576.576 0 0 0-.555-.608h-.054a.587.587 0 1 0 0 1.173ZM30.771 6.684a.598.598 0 0 0 .598-.587.599.599 0 0 0-1.042-.474.598.598 0 0 0-.143.474.587.587 0 0 0 .587.587ZM32.239 7.118a.598.598 0 0 0 .64-.597.596.596 0 0 0-.64-.566.598.598 0 0 0 0 1.185M33.619 7.76a.587.587 0 0 0 .587-.587.598.598 0 0 0-.587-.598.598.598 0 1 0 0 1.184ZM44.15 8.77a.598.598 0 0 0 .597.587.598.598 0 0 0 .587-.598.597.597 0 1 0-1.185 0M44.943 9.618a.598.598 0 0 0-.598.587.598.598 0 0 0 .598.597.598.598 0 0 0 .587-.597.587.587 0 0 0-.587-.587ZM44.65 11.052a.587.587 0 0 0-.588.587.598.598 0 0 0 1.185 0 .598.598 0 0 0-.598-.587ZM41.476 6.542a.587.587 0 1 0 0-1.173.598.598 0 0 0-.598.586.598.598 0 0 0 .598.587ZM42.943 7.042a.598.598 0 0 0 .598-.587.598.598 0 0 0-.598-.597.598.598 0 0 0 0 1.184ZM44.106 8.01a.598.598 0 1 0-.598-.587.587.587 0 0 0 .587.586M39.933 6.684a.598.598 0 0 0 .554-.587.598.598 0 0 0-.598-.598.598.598 0 0 0-.587.598.575.575 0 0 0 .587.587M38.466 7.118a.597.597 0 0 0 .597-.597.598.598 0 0 0-1.012-.394.587.587 0 0 0-.172.415.598.598 0 0 0 .587.598M36.499 7.162a.597.597 0 1 0 1.184 0 .587.587 0 1 0-1.174 0M34.347 10.368h.706V7.836a2.349 2.349 0 0 1-.663-.207l-.043 2.739ZM36.368 7.63c-.213.1-.44.17-.674.206v2.532h.674V7.629ZM34.977 4.869V3.325a1.88 1.88 0 0 0-1.456 1.544h1.456ZM35.738 4.869h1.489a1.891 1.891 0 0 0-1.446-1.51l-.043 1.51ZM33.543 5.564a1.858 1.858 0 0 0 3.63 0h-3.63ZM36.368 3.108a2.076 2.076 0 0 1-.804-1.315 2.326 2.326 0 0 1 1.337.902V.61c-.33.435-.804.74-1.337.858.117-.55.384-1.058.772-1.467h-1.957c.385.41.648.917.761 1.467A2.336 2.336 0 0 1 33.803.61v2.086c.333-.43.806-.73 1.337-.848-.085.527-.374 1-.804 1.315l2.032-.054ZM37.335 46.318a14.857 14.857 0 0 1-3.792-.598 16.785 16.785 0 0 1-3.26-1.271 1.5 1.5 0 0 1-.903.304 1.283 1.283 0 0 1-.608-.142 1.858 1.858 0 0 1-1.087-1.423 1.01 1.01 0 0 1 0-.174v-.054c-.42-.377-.819-.776-1.195-1.196a14.13 14.13 0 0 1-2.663-4.488 15.296 15.296 0 0 1-.685-2.413c.254.098.522.153.794.163a1.349 1.349 0 0 0 1.032-.489c.098 0 .109.337.098.457A1.934 1.934 0 0 0 27 33.58c.11.716.292 1.418.544 2.097a10.128 10.128 0 0 0 1.814 3.174c.262.313.549.604.86.87.181-.065.372-.102.564-.11a1.608 1.608 0 0 1 1.163.576 1.575 1.575 0 0 1 .424 1.261c1.98.902 4.034 1.63 6.14 2.174.94.252 1.907.397 2.88.434.188.011.377.011.565 0a.685.685 0 0 1-.13.12 8.181 8.181 0 0 1-1.24.956 6.868 6.868 0 0 1-3.249 1.185Zm-.913-2.326a.728.728 0 0 0-.728.739.716.716 0 0 0 .728.717.728.728 0 0 0 0-1.456Zm-8.085-3.097c.08.172.17.339.272.5l.054.097-.065.087a3.195 3.195 0 0 0-.554 1.565 1.575 1.575 0 0 0 .836 1.087c.141.074.298.111.457.109.195 0 .386-.053.554-.152-.174-.087-.413-.24-.587-.348l-.098-.065a.64.64 0 0 1-.37-.38.271.271 0 0 1 0-.131c.019-.244.097-.48.23-.685l.053-.087.066-.097.087-.142.152.087c.18.111.386.175.598.185a.271.271 0 0 0 .24-.056.283.283 0 0 0 .075-.096c.108-.196.054-.533-.326-.837l-.13-.109.206-.228c.13-.162.271-.315.424-.456.13-.122.3-.191.478-.196.189.008.37.08.51.207l.196.13.12.098.195.152a1.411 1.411 0 0 0-.337-.717 1.302 1.302 0 0 0-.912-.457c-.794 0-1.63.935-1.729 1.033l-.065.076h-.098l-.532-.174Zm4.597 1.88a.793.793 0 1 0 0 1.587.804.804 0 0 0 .804-.794.794.794 0 0 0-.804-.793ZM35.368 54.403a.88.88 0 0 1-.88-.88 1.01 1.01 0 0 1 .054-.315c-.706 0-.902-.804-1.086-1.37-.087-.337-.153-.597-.294-.597s-.337.293-.478.63a2.37 2.37 0 0 1-.63-1.434 1.272 1.272 0 0 1 1.271-1.272.99.99 0 0 1 .392.065v-2.738a14.25 14.25 0 0 0 3.26.5v2.227c.119-.039.244-.057.37-.054a1.272 1.272 0 0 1 1.27 1.272 2.336 2.336 0 0 1-.63 1.423c-.174-.38-.293-.62-.489-.62-.195 0-.206.261-.282.598-.142.566-.326 1.337-1.087 1.37.017.104.017.21 0 .315a.891.891 0 0 1-.76.88Zm0-1.37a.488.488 0 1 0 .022.978.488.488 0 0 0-.022-.977Zm0-1.651c.24 0 .38.109.38.304a.533.533 0 0 1-.086.261.543.543 0 0 0 .456.87c.468 0 .609-.544.76-1.087.153-.544.24-.935.61-.935a.663.663 0 0 1 .5.391c.129-.242.207-.508.228-.782a.88.88 0 0 0-.87-.88c-.597 0-.76.489-.89.88-.131.391-.175.5-.37.5a.337.337 0 0 1-.37-.305.913.913 0 0 1 .304-.543c.22-.21.36-.49.392-.793 0-.565-.533-1.087-1.087-1.489-.544.445-1.087.924-1.087 1.489.041.302.183.581.402.793a.979.979 0 0 1 .304.544.337.337 0 0 1-.336.336c-.207 0-.283-.228-.37-.5-.087-.271-.293-.88-.891-.88a.88.88 0 0 0-.87.88c.016.275.094.543.229.783a.663.663 0 0 1 .5-.391c.37 0 .489.456.608.934.12.478.283 1.087.761 1.087a.554.554 0 0 0 .609-.489.62.62 0 0 0-.11-.337.445.445 0 0 1-.086-.26c0-.196.12-.305.348-.305l.032-.076Z"
-			/>
-			<path
-				fill="#fff"
-				d="M63.505 51.838a.227.227 0 0 0-.294-.14c-2.608.76-5.836-1.392-8.955-3.479-2.64-1.76-5.075-3.412-7.248-3.412a4.218 4.218 0 0 0-2.087.49 3.75 3.75 0 0 0-1.663 1.955 11.956 11.956 0 0 1-1.032 1.956c-.36.413-.587.924-.652 1.468a1.578 1.578 0 0 0 .543 1.152c.401.38.86.695 1.359.934.13.065.26.142.402.196.51.25.837.424.89.717a17.073 17.073 0 0 1-3.868 1.087c-1.838.315-3.7.468-5.564.456a31.74 31.74 0 0 1-5.554-.456 16.724 16.724 0 0 1-3.89-1.087c0-.424.37-.608.934-.87l.402-.195a6.613 6.613 0 0 0 1.337-.848 1.346 1.346 0 0 0 .544-1.086 2.793 2.793 0 0 0-.576-1.5 12.464 12.464 0 0 1-1.087-1.924 3.685 3.685 0 0 0-1.663-1.912 4.512 4.512 0 0 0-2.087-.5c-2.173 0-4.618 1.652-7.248 3.412-3.12 2.087-6.347 4.238-8.955 3.478a.25.25 0 0 0-.293.141.24.24 0 0 0 .13.304c.778.383 1.508.857 2.173 1.413a5.043 5.043 0 0 0 2.804 1.424 6.782 6.782 0 0 0 3.184-.946 33.383 33.383 0 0 0 3.185-2.01c2.065-1.402 4.184-2.858 6.205-3.021-.217.217-.413.402-.608.576a2.445 2.445 0 0 0-1.087 1.923c.039.703.247 1.385.608 1.989.109.206.218.445.337.717.206.467.36.956.457 1.456.054.24.13.5.206.816a3.063 3.063 0 0 0 1.848 2.173c1.717.87 4.662 1.315 8.792 1.315s7.074-.446 8.791-1.315a3.086 3.086 0 0 0 1.848-2.174c.076-.315.152-.576.196-.815.107-.498.26-.985.456-1.456.163-.37.326-.663.467-.924a3.412 3.412 0 0 0 .49-1.717 2.575 2.575 0 0 0-1.163-1.988c-.207-.185-.413-.37-.642-.598 2.032.152 4.152 1.608 6.206 3.021a34.893 34.893 0 0 0 3.195 2c.95.565 2.026.883 3.13.923a5.042 5.042 0 0 0 2.803-1.423 11.392 11.392 0 0 1 2.174-1.413.25.25 0 0 0 .13-.304M41.737 43.568a13.69 13.69 0 0 1-3.043-.5c-.283-.065-.457-.12-.707-.195l-.239-.066c.023-.282.085-.56.185-.826.085-.27.221-.52.402-.739a.477.477 0 0 0 .163-.13.315.315 0 0 1 .152-.141 8.433 8.433 0 0 0 2.706-2.174 10.107 10.107 0 0 0 1.826-3.162c.407-1.153.627-2.364.652-3.587h-8.118v9.042c.641 0 1.848.152 1.858.152-.302.38-.48.842-.51 1.326a38.348 38.348 0 0 1-2.793-1.021 1.291 1.291 0 0 1 .956-.457v-8.987h-8.303a3.152 3.152 0 0 0-.326-1.087A2.173 2.173 0 0 1 26 29.462a2.848 2.848 0 0 0-.673-1.467 53.74 53.74 0 0 0-1.468-1.63 14.238 14.238 0 0 1 2.663-4.445 12.76 12.76 0 0 1 2.532-2.174 1.815 1.815 0 0 1-.989-1.564 1.457 1.457 0 0 1 .49-1.087 14.203 14.203 0 0 0-2.196-4.619l1.674-.706a2.663 2.663 0 0 0-.37 1.901c.44-.308.75-.77.87-1.293.13.627.334 1.236.608 1.815a2.065 2.065 0 0 0-.847-.239c-.18.007-.357.048-.522.12a1.923 1.923 0 0 0 1.565.76c.18 0 .358-.022.532-.065.283-.065 1.196-.348 1.163-.924a.587.587 0 0 0-.543-.576.337.337 0 0 0-.25.185.88.88 0 0 0 0 .663 1.728 1.728 0 0 1-.62-1.402.522.522 0 0 1 .37-.478h.163c.565 0 1.01.782 1.174 1.282.054-.424.282-1.423.978-1.423a.5.5 0 0 1 .423.304 2.011 2.011 0 0 1-.195 1.543c0-.489-.304-.75-.522-.75-.108 0-.478.087-.434.62.043.532.652.728 1.434.728a1.834 1.834 0 0 0 2.043-1.609 2.5 2.5 0 0 0-1.456.892c.087-.759.087-1.524 0-2.283a2.174 2.174 0 0 0 1.456.935 2.772 2.772 0 0 0-1.206-1.554h3.043a2.74 2.74 0 0 0-1.217 1.554 2.25 2.25 0 0 0 1.456-.891 8.913 8.913 0 0 0 0 2.282 2.5 2.5 0 0 0-1.489-.891 1.848 1.848 0 0 0 2.065 1.576c.782 0 1.39-.305 1.434-.728a.556.556 0 0 0-.108-.446.457.457 0 0 0-.326-.163c-.228 0-.478.25-.522.75a1.967 1.967 0 0 1-.174-1.543.499.499 0 0 1 .424-.305c.685 0 .913.99.978 1.424.196-.62.674-1.282 1.163-1.282h.163a.522.522 0 0 1 .37.478 1.695 1.695 0 0 1-.62 1.38.848.848 0 0 0 0-.652.348.348 0 0 0-.26-.185.565.565 0 0 0-.544.565c0 .587.89.859 1.173.935.175.044.353.066.533.065a1.965 1.965 0 0 0 1.565-.772 1.195 1.195 0 0 0-.51-.108 2.053 2.053 0 0 0-.849.228c.283-.576.49-1.186.62-1.815.133.524.464.976.924 1.26a2.598 2.598 0 0 0-.37-1.88l1.674.707a14.498 14.498 0 0 0-2.25 4.554 1.455 1.455 0 0 1 .522 1.086 1.847 1.847 0 0 1-1 1.565 12.395 12.395 0 0 1 2.521 2.174 12.14 12.14 0 0 1 1.576 2.173 1.564 1.564 0 0 0-.239.294c-.09.136-.191.263-.304.38l-.174.196a1.087 1.087 0 0 0-.261.63c.012.272.098.535.25.76.196.327 1.489 2.87 1.804 3.511.227.458.558.857.967 1.163.01.18.01.362 0 .543v.261a15.443 15.443 0 0 1-.978 5.434 14.573 14.573 0 0 1-2.652 4.63 13.96 13.96 0 0 1-1.804 1.684 1.086 1.086 0 0 1-.684.065Zm-6.02-11.954h8.128a11.456 11.456 0 0 0-.674-3.717 9.987 9.987 0 0 0-1.815-3.162 8.499 8.499 0 0 0-2.706-2.174 7.433 7.433 0 0 0-2.934-.76v9.813Zm-.49-9.781a7.357 7.357 0 0 0-3.14.772 8.575 8.575 0 0 0-2.696 2.173 10.226 10.226 0 0 0-1.826 3.163 11.387 11.387 0 0 0-.663 3.716h8.314l.011-9.824Zm.152-4.064c2.04-.066 4.053.48 5.782 1.565.64-.38.923-.75.923-1.196 0-.685-.836-1.174-2.51-1.445a27.1 27.1 0 0 0-4.195-.261 27.165 27.165 0 0 0-4.206.26c-1.662.283-2.51.761-2.51 1.446 0 .446.283.815.913 1.195a10.335 10.335 0 0 1 5.803-1.564Z"
-			/>
-			<path
-				fill="#fff"
-				d="M39.4 31.136a.533.533 0 0 1-.141-.533h-2.402a.14.14 0 0 0 0 .109.13.13 0 0 0 .109 0 .228.228 0 0 0 .152-.076.315.315 0 0 1-.282.326h-.077a.228.228 0 0 1-.12-.261.576.576 0 0 1-.369.141c0-.134.046-.265.13-.37a.217.217 0 0 1-.217-.119.261.261 0 0 1 0-.217.36.36 0 0 1 .283-.141.195.195 0 0 0-.076.206.174.174 0 0 0 .152.098v-3.434h-.152a.554.554 0 0 1-.38-.163c.099-.098.23-.155.369-.163.054-.01.11-.01.163 0v-3.13a.152.152 0 0 0-.076.282.315.315 0 0 1-.272-.152.282.282 0 0 1 0-.228.293.293 0 0 1 .218-.098.609.609 0 0 1-.11-.391.564.564 0 0 1 .37.196.207.207 0 0 1 .142-.196h.065c.07.004.137.035.185.087a.325.325 0 0 1 .087.26.196.196 0 0 0-.163-.097.152.152 0 0 0-.152.13 6.685 6.685 0 0 1 1.956.902 7.955 7.955 0 0 1 1.771 1.641c.052.084.11.164.174.24a.511.511 0 0 1 .456-.24h.066a.533.533 0 0 1-.337.435 9.226 9.226 0 0 1 1.565 4.14.152.152 0 0 0 .152-.108.175.175 0 0 0-.076-.196.304.304 0 0 1 .271.13.24.24 0 0 1 0 .207.218.218 0 0 1-.228.141.554.554 0 0 1 .141.36.543.543 0 0 1-.358-.131.24.24 0 0 1-.13.26h-.077a.304.304 0 0 1-.282-.315.184.184 0 0 0 .141.076.173.173 0 0 0 .12 0 .119.119 0 0 0 0-.108h-2.63a.532.532 0 0 1-.13.521v-.021Zm.359-1.087h.065a.26.26 0 0 1 .141.282h2.033a.478.478 0 0 1-.163-.39.576.576 0 0 1 .293.086 9.063 9.063 0 0 0-1.217-3.26.207.207 0 0 1-.109 0 .152.152 0 0 1-.098 0 .207.207 0 0 1-.087-.26.805.805 0 0 1-.315.086h-.076a.532.532 0 0 1 .196-.348.25.25 0 0 1-.24-.163.228.228 0 0 1 .077-.228 6.933 6.933 0 0 0-1.609-1.478 6.52 6.52 0 0 0-1.565-.76.598.598 0 0 1 .087.336.62.62 0 0 1-.304-.12v2.37a.207.207 0 0 1 .217.13.228.228 0 0 1-.097.283.511.511 0 0 1 .358.152.49.49 0 0 1-.358.163c.13.109.13.206.097.272a.206.206 0 0 1-.217.13v2.826a.51.51 0 0 1 .348-.13.564.564 0 0 1-.12.347h1.728a.25.25 0 0 1 .142-.272h.065a.282.282 0 0 1 .206.109.478.478 0 0 1 .163-.348c.09.095.143.218.152.348a.292.292 0 0 1 .207-.12v-.043Z"
-			/>
-			<path
-				fill="#fff"
-				d="M41.128 29.136a.445.445 0 0 0 .228-.196.282.282 0 0 1-.228 0c.076-.054.174 0 .076-.152-.098-.152-.163-.152-.26 0a.565.565 0 0 1-.707-.391.217.217 0 0 1 0-.12.392.392 0 0 0 .217.174h.163c.06.022.125.022.185 0a.283.283 0 0 0 .25-.239c0-.271-.228-.391-.51-.63l-.261-.24a2.326 2.326 0 0 0-.272-.705 1.657 1.657 0 0 0-.076-.316.892.892 0 0 0 .141 1.022c.272.293.858.598.826.837a.162.162 0 0 1-.141.13.12.12 0 0 1-.087 0l-.185-.087a.934.934 0 0 1-.196-.152 4.819 4.819 0 0 1-.597-.739 1.087 1.087 0 0 1-.109-.945c.185-.272-.12-.805-.294-.924a.5.5 0 0 0 0-.283.271.271 0 0 1-.162 0c0-.087-.066-.108-.13-.12-.066-.01-.11-.086-.131-.13-.022-.043-.087.12-.163.12h-.163a.272.272 0 0 1-.196.065c-.12 0-.109.065-.109.141a.174.174 0 0 0 .076.261l.076.098.066-.109s.184-.065.217 0c.033.066 0 .087-.065.098h-.228v.13c0 .044.065.098.195.066.13-.033.359-.055.359.054 0 .293-.326.283-.402.62a.534.534 0 0 1-.196-.185 1.228 1.228 0 0 0-.26-.315 1.353 1.353 0 0 0-.24-.196c-.108-.065-.12-.12-.12-.25v-.283a.205.205 0 0 0 0-.087.174.174 0 0 0-.162-.097c.031.025.054.059.065.098a.076.076 0 0 0 0 .054.066.066 0 0 0-.016.076c.003.008.009.015.016.022v.065s0-.163-.13-.174c0-.076-.077-.13-.174 0h.076a.184.184 0 0 0 0 .076c0 .065 0 .109.12.109.119 0-.185 0-.185.065l-.098.065s-.055.098 0 .152a.13.13 0 0 1 0-.087h.065s-.076.152 0 .185a.11.11 0 0 1-.127.037.108.108 0 0 1-.069-.113c-.108.174.076.293.229.228.152-.065 0 .196 0 .196a.076.076 0 0 0 .076.098c.02.045.046.089.076.13.115.263.313.48.565.62a1.345 1.345 0 0 0-.402.38c-.098.087-.229.195-.326.141-.098-.054 0-.065-.098-.109h-.174a.163.163 0 0 0-.185.066.163.163 0 0 1 .142 0v.065s0 .076.206.065a.11.11 0 0 1-.098 0c-.087 0-.12-.087-.185 0-.065 0-.13.163-.054.217a.153.153 0 0 1 .054-.108c0 .076.098.12.218-.055.12-.174-.098.207 0 .218h.054a.163.163 0 0 0 .196.098.25.25 0 0 1-.109-.13h.065c.055-.066 0-.142 0-.196a.12.12 0 0 0 .13.087.163.163 0 0 0 .066-.11.555.555 0 0 1 .771.099c0-.066 0-.152.185-.185a.207.207 0 0 1 .228 0 .468.468 0 0 1 .12-.218c0 .5.674.5.826.794a1.228 1.228 0 0 0-.554-.109c-.163 0-.468.174-.5.37a.707.707 0 0 0 .293.543h-.445c-.153 0-.174-.26-.359-.185-.065 0-.163.087-.152.163a.23.23 0 0 1 .108-.065s0 .065.077.065h.076c0 .066-.229 0-.185.152v.055c0 .065 0 .12.152.087a.055.055 0 0 1 0-.076.153.153 0 0 0 0 .065c0 .076.098.141.196.076a.12.12 0 0 1-.098-.054.075.075 0 0 0 0-.066.109.109 0 0 0 .141.055.076.076 0 0 0 .065 0h.087c0-.076 0-.13.196 0l.109-.087s.14.065.206 0c.065-.065 0-.087 0-.174s-.109-.13-.13-.283c-.022-.152.293-.39.5-.347a1.153 1.153 0 0 0 .64.706c.229.076.327-.174.414-.141.087.032 0 0 0 .065.011.09.011.181 0 .272 0 .065-.152 0-.24 0h-.086a.185.185 0 0 0-.163.163.402.402 0 0 1 .13-.076.109.109 0 0 0 .109.108.098.098 0 0 0 .087 0s-.174.142-.076.185a.11.11 0 0 0 .015.084.107.107 0 0 0 .072.046v-.12c.054 0 0-.13.097-.162.098-.033 0 .163 0 .174a.107.107 0 0 0 .093.051.109.109 0 0 0 .092-.051h-.087v-.055c0-.152-.087-.163-.087-.217 0-.054.152 0 .152-.174.076 0 .142-.087.153-.141a.36.36 0 0 1-.185 0l.543.152ZM33.032 39.993a.544.544 0 0 1-.576-.511.424.424 0 0 1 .434-.402.238.238 0 0 1 .207.065.151.151 0 0 1 0 .152c0 .055-.076.065-.152.065s-.163 0-.163.13c0 .131.087.175.25.175a.326.326 0 0 0 .348-.261v-.087c0-.359-.294-.413-.468-.413a.813.813 0 0 0-.282 0 5.185 5.185 0 0 1-3.739-5.097h-.163l.163-.38a.337.337 0 0 1-.25.12.315.315 0 1 1 .12-.607.306.306 0 0 1 .185.291l.163-.38a3.369 3.369 0 0 0 1.575.5c.66-.024 1.31-.16 1.924-.402.226-.097.463-.163.706-.196a.816.816 0 0 1 .598.228.837.837 0 0 1 .207.63v.12h.14a.272.272 0 0 1 .262.272.26.26 0 0 1-.272.239h-.141l-.109 4.347h.141v.391c0 .13-.195.141-.315.141a.718.718 0 0 1-.152.576.816.816 0 0 1-.641.294Zm-3.26-5.934a3.794 3.794 0 0 0 .836 2.804 5.325 5.325 0 0 0 2.174 1.576l-3.01-4.38Zm.358 0 2.75 3.967c0-.185.065-.424.097-.63v-.185l-2.173-3.185h-.424l-.25.033Zm1.663-.293a4.97 4.97 0 0 1-.62.184l1.957 2.804c.054-.326.097-.597.14-.89l-1.477-2.098Zm1.576-.642a3.042 3.042 0 0 0-.87.294l-.424.174 1.24 1.782.054-.348c0-.358.098-.62.12-.782v-.109H33.4a.294.294 0 0 1-.206-.087.261.261 0 0 1 0-.358.25.25 0 0 1 .206-.076h.152v-.446a.185.185 0 0 0-.184-.044ZM30.5 25.898a.13.13 0 0 1-.109-.055.086.086 0 0 1 0-.087v-.141a.173.173 0 0 1-.12 0c-.054 0-.065 0-.076-.098h-.076v-.065a.152.152 0 0 1 .12-.076h.054c-.108 0-.141-.055-.152-.076a.108.108 0 0 0-.087 0h.054a.076.076 0 0 1 .087 0h.066a.086.086 0 0 1 .076 0 .269.269 0 0 1 0 .076.108.108 0 0 0 .108 0 .64.64 0 0 0 .283-.076c.05-.01.102-.01.152 0h.141c.127-.021.252-.058.37-.109a.489.489 0 0 1-.109-.152.076.076 0 0 0 0 .054v-.054h-.087a.14.14 0 0 1-.152-.065.12.12 0 0 1 0-.142.075.075 0 0 0 0 .076.075.075 0 0 0 .065 0 .108.108 0 0 0 .109-.076c-.054 0-.076 0-.087-.065a.14.14 0 0 1 .076-.098h-.12v.066c0 .065 0 0-.054-.066a.087.087 0 0 1 0-.065.086.086 0 0 1 .04-.056.088.088 0 0 1 .069-.009h.076a.087.087 0 0 1-.098 0v-.054h-.065a.12.12 0 0 1 .087 0 .096.096 0 0 1 0 .054c.054 0 .065.098.076.163v-.065a.787.787 0 0 1 0-.087.087.087 0 0 0 0-.065.12.12 0 0 1 .109.087v.076a.412.412 0 0 1 0 .108.392.392 0 0 0 0 .12c0 .087 0 .141.076.195l.13.098a2.4 2.4 0 0 1 .26-.739s.066-.054.088-.108c.022-.055 0-.055 0-.087v-.066s.054 0 .076-.065.065-.065.087-.065a.25.25 0 0 1 0-.196c.039-.01.08-.01.12 0h.173a.218.218 0 0 1 .142 0c.06.02.114.058.152.11v.053a.076.076 0 0 1 0 .087.26.26 0 0 1 .12.066.172.172 0 0 1-.11.12c.055.053 0 .151-.054.25a.293.293 0 0 0-.076.151c.137-.016.276-.016.413 0 .162.009.322.034.478.076.11.034.223.055.337.066.446 0 .74-.11.74-.272 0-.163-.36-.228-.74-.26-.38-.034-.913 0-.913-.381a.326.326 0 0 1 .174-.25.164.164 0 0 1 0-.12c.098-.095.212-.173.337-.228a.815.815 0 0 0 .304-.196.37.37 0 0 1-.217.337.5.5 0 0 0 .304-.185.826.826 0 0 1 .652-.304h.153c-.217.05-.411.168-.555.337-.225.24-.53.392-.858.424a.228.228 0 0 0-.12.174c0 .174.304.195.663.217.359.022.967.065.967.435 0 .185-.282.38-.89.38.099.012.197.033.293.066h.315v.054a.152.152 0 0 0 .098-.055.13.13 0 0 1 0 .142v.13a.066.066 0 0 1 .054.065.087.087 0 0 1-.04.098.086.086 0 0 1-.036.011v.054a.108.108 0 0 1 0 .185h.065-.054a.076.076 0 0 1-.076 0v-.13a.13.13 0 0 0-.055.076v.141a.076.076 0 0 1-.076-.098s0-.087.087-.108a.077.077 0 0 1-.065 0 .144.144 0 0 1-.161.108.144.144 0 0 1-.122-.152h-.087c0-.065.066-.109.12-.12a.108.108 0 0 1 .076 0h.087a.055.055 0 0 0 .027.008c.01 0 .02-.003.027-.007v-.066a.217.217 0 0 1 0-.12h-.12a.358.358 0 0 0-.151 0h-.098a.64.64 0 0 1-.315-.075.388.388 0 0 1 0 .108v.066a.262.262 0 0 1 .076.25h-.294a.054.054 0 0 0 0 .065.076.076 0 0 1 0 .087.097.097 0 0 1-.087-.055h-.086v.055h.086a.109.109 0 0 1-.152 0v-.055a.077.077 0 0 1-.065 0 .12.12 0 0 1 0-.087.119.119 0 0 1 .087 0 .152.152 0 0 0-.098.055.097.097 0 0 1 0-.076.185.185 0 0 1 .109-.066c.054 0 .087.055.12.098.032.044 0 .087.097.087h.142a.423.423 0 0 0-.163-.065.076.076 0 0 1-.066 0 .316.316 0 0 1 0-.26.434.434 0 0 1-.195.075h-.152a.26.26 0 0 0-.174 0 .1.1 0 0 0 0 .076.12.12 0 0 1 0 .087.195.195 0 0 1-.109 0l-.065.055a.512.512 0 0 1-.272.12v-.088a.153.153 0 0 1-.141.076s-.12.13-.174.13a.086.086 0 0 0 .054-.075.337.337 0 0 1-.13.054c-.076 0-.13 0-.141.174a.152.152 0 0 1 0-.098v-.098h-.185a.076.076 0 0 0 0-.054.207.207 0 0 1-.141.152.13.13 0 0 0 0-.152.13.13 0 0 1-.055.054.12.12 0 0 0 0-.098.163.163 0 0 1-.13.13c-.076 0-.098-.097-.098-.119a.153.153 0 0 0 0 .066.076.076 0 0 1-.054.087.131.131 0 0 0 0 .119l-.576-.12Zm1.64-1.663s.055 0 0 .076c-.054.076 0 0-.065 0s0 0 0 .065a.087.087 0 0 0 0 .087.098.098 0 0 0 .076-.065v-.131h.055v-.065l-.163-.076h-.065.054l.109.109ZM29.326 28.527a.152.152 0 0 1-.109-.076.098.098 0 0 1 0-.108.152.152 0 0 1 0-.098l.055-.065a.38.38 0 0 1-.152.065c-.055 0-.077 0-.088-.098a.118.118 0 0 0-.097.076.076.076 0 0 1 0-.098.152.152 0 0 1 .13-.076.075.075 0 0 1 .065 0 .077.077 0 0 0 .066 0 .158.158 0 0 1-.164-.12.163.163 0 0 0-.097-.064H29a.13.13 0 0 1 .109 0h.065c.04.008.074.032.098.065a.333.333 0 0 1 0 .087v.054a.14.14 0 0 0 .13 0 .934.934 0 0 0 .326-.098.577.577 0 0 1 .24 0h.173a.326.326 0 0 1 0-.195.065.065 0 0 1-.054 0 .151.151 0 0 1-.109-.055l-.087-.054a.86.86 0 0 1-.185-.152s0-.055-.054-.055a.097.097 0 0 0 0 .055c-.054 0-.065-.055-.065-.098a.076.076 0 0 0 0 .054.108.108 0 0 1 0-.076.207.207 0 0 1-.109 0 .173.173 0 0 1-.152-.087.164.164 0 0 1 0-.184.11.11 0 0 0 0 .108.087.087 0 0 0 .076 0 .11.11 0 0 0 .109-.109.087.087 0 0 1-.087-.054.12.12 0 0 1 .054-.13h-.13v.076a.14.14 0 0 1 0-.076.13.13 0 0 1 0-.087.076.076 0 0 1 .087-.076h.087a.219.219 0 0 1-.12 0 .088.088 0 0 1 0-.076.216.216 0 0 1 0-.076h-.065a.144.144 0 0 1 .098-.055.087.087 0 0 1 .076.087c.054 0 .087.098.108.174v-.076a.065.065 0 0 1 0-.087.098.098 0 0 0-.054-.109.153.153 0 0 1 .141.11v.086a.547.547 0 0 1 0 .13.423.423 0 0 0 0 .153c0 .12 0 .173.109.239.075.042.14.097.196.163h.239v-.229a.206.206 0 0 1-.087-.087s.065-.054.087-.13l.054-.109-.065-.065s.065 0 .098-.087c.032-.087.065-.076.097-.076a.305.305 0 0 1 0-.217c.037.009.071.027.098.054h.358a.488.488 0 0 1 .153.109v.152c.055.002.108.02.152.054a.163.163 0 0 1-.12.142c0 .076 0 .206-.086.315a.597.597 0 0 1-.055.087c.233-.043.47-.065.707-.066.413 0 .858 0 1.206.077.192.012.384.012.576 0 .304 0 .75 0 .75-.163s-.087-.153-.598-.153h-.478a5.3 5.3 0 0 1-.826.066 1.381 1.381 0 0 1-.522-.066.293.293 0 0 1-.217-.271c0-.229.24-.261.5-.261h.304c.1-.053.213-.08.326-.076h.468a.392.392 0 0 0 .26-.065c0 .087-.108.195-.467.195h.228c.134.01.269.01.402 0a1.45 1.45 0 0 1 .38 0 .772.772 0 0 1 .523.174.701.701 0 0 0-.218 0c-.19 0-.38.022-.565.065h-.054c-.241.042-.484.067-.728.076-.173 0-.344-.021-.511-.065a.206.206 0 0 1-.13-.076H32c-.153 0-.316 0-.316.141 0 .142.174.174.555.174h1.336c.522 0 .783.109.783.305 0 .195-.544.315-.946.315h.674a.065.065 0 0 1 .065.054h.054a.13.13 0 0 0 .11 0v.152s0 .087.064.087a.088.088 0 0 1-.076.055.088.088 0 0 1 .066.087.109.109 0 0 1-.087.119v.054a.13.13 0 0 1 0 .152.23.23 0 0 1 0 .066h.065-.065a.109.109 0 0 1-.098-.066c-.065 0 0-.119 0-.173a.12.12 0 0 0-.055.097v.055a.14.14 0 0 0 0 .13.152.152 0 0 1-.087-.141s.055-.098.11-.13a.108.108 0 0 1-.077 0 .089.089 0 0 1-.069-.022.085.085 0 0 1-.029-.066.25.25 0 0 0-.13.066.185.185 0 0 1 .141-.152.12.12 0 0 1 .087 0h.217a.163.163 0 0 0 0-.076.218.218 0 0 1 0-.142s0-.054-.054-.065h-.272a.283.283 0 0 0-.163 0 .684.684 0 0 1-.456.109.78.78 0 0 1-.315-.066h-.054a.153.153 0 0 1 0 .12v.065a.293.293 0 0 1 .086.294h-.358v.076a.087.087 0 0 1-.065.098.131.131 0 0 1-.11-.066.205.205 0 0 1 0 .087h.099a.12.12 0 0 1-.087 0 .088.088 0 0 1-.087 0 .153.153 0 0 1 0-.076v-.054H32.4a.087.087 0 0 0 .055.065.12.12 0 0 1-.087 0 .066.066 0 0 1 0-.065.13.13 0 0 1 0-.141h.174-.076a.262.262 0 0 0-.12.065.175.175 0 0 1 0-.098.186.186 0 0 1 .13-.065h.055c.054 0 .087.054.13.108.044.055.066.098.109.098h.065c.05-.01.102-.01.152 0a.337.337 0 0 0-.195-.076.097.097 0 0 1-.066-.065.326.326 0 0 1 0-.283.522.522 0 0 1-.195.087h-.272a.392.392 0 0 0-.358.174.283.283 0 0 1-.283.152h-.185a.674.674 0 0 1-.456.348.174.174 0 0 0 .065-.087.423.423 0 0 1-.152 0c-.098 0-.152 0-.174.207a.229.229 0 0 1 0-.12.152.152 0 0 1 0-.098H30.5a.26.26 0 0 1-.142.163.174.174 0 0 0 0-.184.13.13 0 0 1-.076.076.109.109 0 0 0 0-.109.152.152 0 0 1-.141.141c-.087 0-.109-.087-.12-.13-.01-.044 0 0 0 .054 0 .055 0 .098-.076.109a.163.163 0 0 0 0 .141l-.62-.348Zm1.38-1.782h-.054s.076 0 .054.076 0 0-.076 0 0 0 0 .055a.303.303 0 0 0 0 .097.109.109 0 0 0 .054 0 .12.12 0 0 0 .098-.065v-.185h.065v-.065l-.14.033v.054Zm.261-.13a.152.152 0 0 0 0 .13h.054a.195.195 0 0 1 .098 0 .727.727 0 0 1-.228-.098l.076-.032Zm-.337-.196c.011.035.011.073 0 .109v.054l.076.022s.065 0 .087-.065l-.163-.12Z"
-			/>
-			<path
-				fill="#fff"
-				d="M32.239 30.95a.098.098 0 0 1-.087 0 .69.69 0 0 1 0-.075h-.055s0 .054.054.065a.097.097 0 0 1-.076 0 .087.087 0 0 1 0-.054.11.11 0 0 1 0-.098.13.13 0 0 1 .098 0h.066-.077a.262.262 0 0 0-.12.065.12.12 0 0 1 0-.087.152.152 0 0 1 .131-.065c.066 0 .098.054.13.109 0 .054.066.097.12.097h.185a.391.391 0 0 0-.185-.076.119.119 0 0 1-.076-.065.284.284 0 0 1 0-.228.152.152 0 0 1-.098 0h-.239a.75.75 0 0 0-.543.163.567.567 0 0 1-.424.152h-.337s-.065.054-.12.054H30.5l-.25.11a.412.412 0 0 1-.174.064.13.13 0 0 0 .065-.087.467.467 0 0 1-.206 0c-.13 0-.25 0-.261.174a.11.11 0 0 1 0-.108.163.163 0 0 1 0-.098h-.196a.25.25 0 0 1-.163.163.153.153 0 0 0 0-.174.108.108 0 0 1-.076.076v-.098a.12.12 0 0 1 0 .087.118.118 0 0 1-.098.055.11.11 0 0 1-.08-.036.108.108 0 0 1-.029-.084.099.099 0 0 0 0 .065.108.108 0 0 1-.054.098h-.445a.14.14 0 0 0 .054.141.155.155 0 1 1-.054-.304.13.13 0 0 1 0-.098.163.163 0 0 1 .054-.087.46.46 0 0 0 0-.076.336.336 0 0 1-.141.065c-.055 0-.066 0-.076-.098v-.054a.141.141 0 0 0-.098.076.152.152 0 0 1 0-.098.142.142 0 0 1 .109-.065h.065v.055h.098s-.142-.087-.153-.11v-.053a.14.14 0 0 0-.097-.077h.065a.108.108 0 0 1 .108 0h.066a.107.107 0 0 1 .076.066v.076a.196.196 0 0 0 .13.054.718.718 0 0 0 .304-.098c.105-.01.21-.01.316 0h.26v-.065a.456.456 0 0 1 0-.163.152.152 0 0 1-.108.054.794.794 0 0 1-.316-.282h-.054a.065.065 0 0 0 0 .065.108.108 0 0 1-.087-.109H29a.207.207 0 0 1-.109 0 .196.196 0 0 1-.163-.087.174.174 0 0 1 0-.163.163.163 0 0 0 0 .098.098.098 0 0 0 .076 0 .097.097 0 0 0 .098-.087.109.109 0 0 1-.087-.065.13.13 0 0 1 0-.12h-.076v.066a.098.098 0 0 1-.054-.087v-.076a.087.087 0 0 1 .087-.076h.098c-.077 0-.13 0-.12-.109a.184.184 0 0 1 0-.076.214.214 0 0 1-.217-.37.076.076 0 0 1 .076.087c.054 0 .097.109.12.174v-.065a.087.087 0 0 1 0-.054v-.066a.13.13 0 0 0 0-.097.152.152 0 0 1 .13.098v.054a.294.294 0 0 1 0 .141.537.537 0 0 0 0 .13c0 .12 0 .164.108.229a.692.692 0 0 1 .163.141c.055.054.087.098.142.098h.087c.028.01.058.01.086 0a.953.953 0 0 1 0-.174.13.13 0 0 1 0-.054s.066-.066.087-.13l.055-.099-.055-.076.087-.076c.087-.076.055-.065.087-.065a.271.271 0 0 1 0-.206c.045.01.087.033.12.065 0-.055.065-.055.098-.055h.13a.46.46 0 0 1 .13 0l.142.076v.066s.054.065 0 .098a.25.25 0 0 1 .13.065.174.174 0 0 1-.12.141c.066.076 0 .185-.065.293l-.065.11c.295-.077.598-.11.902-.099.5 0 1.087 0 1.49.087h.564c.25 0 .728 0 .728-.163s-.087-.141-.565-.141h-.467c-.343.04-.688.061-1.032.065a2.073 2.073 0 0 1-.63-.065c-.175-.054-.273-.141-.273-.26 0-.12.316-.25.653-.25h.521a.532.532 0 0 1 .315-.066h.642c.093.01.188-.01.271-.054 0 .097-.348.152-.663.152h.207c.143-.002.285-.02.424-.055.13-.01.26-.01.39 0a.576.576 0 0 1 .436.152 1.095 1.095 0 0 0-.261 0 2.646 2.646 0 0 0-.587.088 3.042 3.042 0 0 1-.696.086c-.13.022-.261.022-.39 0a.173.173 0 0 1-.11-.087h-.282c-.196 0-.413 0-.413.13 0 .131.239.175.717.175h1.543c.5 0 .74.087.74.271 0 .185-.63.283-.903.305h.63a.065.065 0 0 1 .077 0h.054a.131.131 0 0 0 .098 0 .14.14 0 0 1 0 .141s0 .087.065.087h-.065c.03.021.053.052.065.087a.109.109 0 0 1-.076.12v.054a.12.12 0 0 1 0 .152v.054a.087.087 0 0 1-.087-.054v-.163a.163.163 0 0 0 0 .098v.163a.109.109 0 0 1-.087-.12.195.195 0 0 1 .098-.12.1.1 0 0 1-.076 0 .12.12 0 0 1-.098-.097H34a.163.163 0 0 1 .13-.141.13.13 0 0 1 .087 0h.152a.163.163 0 0 0 0-.077.262.262 0 0 1 0-.13v-.065h-.119a.36.36 0 0 0-.185 0 .63.63 0 0 1-.434.109.74.74 0 0 1-.305-.076c.01.035.01.073 0 .108a.087.087 0 0 0 0 .087.26.26 0 0 1 .087.26.206.206 0 0 1-.152 0h-.185v.066a.087.087 0 0 1 0 .098s-.076 0-.087-.065h-.087a.077.077 0 0 1 0 .065h-.054.098a.41.41 0 0 1-.706.413Zm-2.228-1.63h-.055s.066 0 0 .088c-.065.087 0 0-.076 0a.12.12 0 0 0 0 .152.12.12 0 0 0 .098-.076V29.32h.065v-.065l-.174-.076h-.065l.207.14ZM40.16 34.7a.097.097 0 0 1-.075 0v-.065.065h-.054a.065.065 0 0 1-.087-.065.12.12 0 0 1 0-.087.163.163 0 0 1 .108 0h.066c-.066 0-.077 0-.077-.054a.184.184 0 0 0-.108.065.175.175 0 0 1 0-.098.218.218 0 0 1 .12-.065c.065 0 .097.065.13.109 0 .054.065.097.108.097a.075.075 0 0 0 .066 0l.13-.39a.393.393 0 0 0-.185-.077l-.076-.065a.392.392 0 0 1 0-.272s0 .065-.087.065h-.25a.729.729 0 0 0-.532.163.532.532 0 0 1-.413.153 1.8 1.8 0 0 1-.24 0 .152.152 0 0 1-.119.065h-.152s-.174.206-.26.206a.133.133 0 0 0 .064-.087.304.304 0 0 1-.14 0c-.088 0-.153 0-.164.196a.163.163 0 0 1 0-.109.098.098 0 0 1 0-.098h-.206a.272.272 0 0 1-.152.174.185.185 0 0 0 0-.184.142.142 0 0 1-.066.076.086.086 0 0 0 0-.098.097.097 0 0 1 0 .076.152.152 0 0 1-.108 0c-.087 0-.11-.076-.11-.13v.065a.11.11 0 0 1 0 .097.142.142 0 0 0 0 .142.13.13 0 0 1-.097 0 .13.13 0 0 1 0-.109.14.14 0 0 1 0-.076v-.065a.315.315 0 0 1-.141 0s-.065 0-.076-.098a.119.119 0 0 0-.087.087.076.076 0 0 1 0-.098.152.152 0 0 1 .12-.076h.13v.055h.087a.305.305 0 0 1-.152-.098v-.055a.14.14 0 0 0-.098-.076h.065a.185.185 0 0 1 .109 0h.054a.13.13 0 0 1 .087.065v.076s.065.055.12.055a.771.771 0 0 0 .304-.098.294.294 0 0 1 .141 0h.185a.282.282 0 0 1-.076-.185.141.141 0 0 1-.076-.098h-.142a.815.815 0 0 1-.293-.282.098.098 0 0 1-.076-.098.13.13 0 0 1-.109 0 .163.163 0 0 1-.177-.163c0-.03.009-.06.025-.087a.153.153 0 0 0 0 .098.12.12 0 0 0 .087 0 .098.098 0 0 0 .098-.098.098.098 0 0 1-.087-.065.12.12 0 0 1 0-.12.097.097 0 0 1-.065 0v.077a.108.108 0 0 1 0-.087.066.066 0 0 1 0-.066v-.054a.088.088 0 0 1 .087-.076h.087a.108.108 0 0 1-.11 0 .087.087 0 0 1 0-.076v-.065h-.064a.098.098 0 0 1 .087 0 .076.076 0 0 1 .076.076s.097.108.108.174a.098.098 0 0 0 0-.055.076.076 0 0 1 0-.097.13.13 0 0 0-.054-.11.184.184 0 0 1 .13.11v.075a.557.557 0 0 1 0 .142.537.537 0 0 0 0 .13c0 .109 0 .174.11.228l.173.142c.04.04.089.074.141.097h.261v-.097a.144.144 0 0 1 0-.066l-.706-.239s.065-.054.098-.12a.444.444 0 0 1 0-.097l-.066-.076c-.065-.076 0 0 .087-.065s.065-.076.098-.076a.239.239 0 0 1 0-.196.174.174 0 0 1 .13 0h.163c.054-.01.11-.01.163 0a.61.61 0 0 1 .163.109v.163c.053.008.102.03.142.065a.174.174 0 0 1-.12.141c0 .066 0 .174-.065.272a.368.368 0 0 0-.065.13c.301-.076.613-.11.924-.097.5 0 1.086 0 1.456.076h.587c.108 0 .728 0 .728-.153 0-.152-.098-.14-.576-.14H39.77a2.042 2.042 0 0 1-.63-.066c-.185 0-.272-.13-.272-.25s.337-.25.663-.25h.554a.784.784 0 0 1 .315-.065h.641a.488.488 0 0 0 .272 0c0 .087-.37.152-.663.152a.74.74 0 0 0 .228 0c.147-.003.292-.02.435-.054.138-.036.28-.054.424-.055.122.002.243.024.358.066a.422.422 0 0 0-.163 0 2.247 2.247 0 0 0-.597.12h-.055a3.255 3.255 0 0 1-.782.13 1.01 1.01 0 0 1-.359 0s-.098 0-.108-.076h-.63c-.262 0-.381 0-.381.13s.25.174.739.174h1.521c.5 0 .74.087.74.282 0 .196-.61.25-.903.272h.674c.076 0 .076 0 .076.065v.055a.184.184 0 0 0 .109-.066.163.163 0 0 1 0 .152.076.076 0 0 0 0 .066.065.065 0 0 1-.076 0 .11.11 0 0 1 0 .087.13.13 0 0 1-.098.108v.055a.119.119 0 0 1 0 .14h-.065a.087.087 0 0 1-.087-.064.207.207 0 0 1 0-.142.153.153 0 0 0-.054.087v.066a.12.12 0 0 0 0 .108.13.13 0 0 1-.077-.12s0-.086.098-.108a.097.097 0 0 1-.076 0 .099.099 0 0 1 0-.065h-.087c0-.065.076-.12.142-.141h.25a.14.14 0 0 0 0-.066.538.538 0 0 1 0-.13v-.076h-.131a.357.357 0 0 0-.185 0 .848.848 0 0 1-.456.098.957.957 0 0 1-.348-.076c.01.035.01.073 0 .108v.076a.24.24 0 0 1 .087.142.141.141 0 0 1 0 .13H40.487v.065a.076.076 0 0 1-.054.087.098.098 0 0 1-.087 0h-.087a.118.118 0 0 1 0 .065h.098a.234.234 0 0 1-.196.424Zm-2.238-1.51v.054a.109.109 0 0 0 0 .13.119.119 0 0 0 .098-.065.24.24 0 0 0 0-.087v-.098h.065a.163.163 0 0 1 0-.065l-.184-.087h.076s.065 0 .054.076l-.109.142Zm.272-.305a.131.131 0 0 0 0 .109.054.054 0 0 0 .054 0 .131.131 0 0 1 .087 0l-.141-.109ZM36.65 37.07a.207.207 0 0 1-.108-.055.175.175 0 0 1 0-.098.22.22 0 0 1 0-.087l.054-.065a.338.338 0 0 1-.14.065c-.055 0-.066 0-.077-.097-.01-.098-.065 0-.087.065a.12.12 0 0 1 0-.087.163.163 0 0 1 .13-.076h.153c-.109 0-.142-.087-.142-.109a.088.088 0 0 0-.087-.065h.066a.108.108 0 0 1 .097 0h.077a.173.173 0 0 1 .086.065.15.15 0 0 1 0 .065.142.142 0 0 0 .12 0 .857.857 0 0 0 .282-.087.272.272 0 0 1 .131 0h.174a.12.12 0 0 0 0-.087v-.163h-.12a.12.12 0 0 1-.076-.065.26.26 0 0 0-.152-.054h-.065a.14.14 0 0 0 0 .065.109.109 0 0 1-.065-.109v-.065a.217.217 0 0 1-.12 0 .185.185 0 0 1-.141-.076.174.174 0 0 1 0-.163.12.12 0 0 0 0 .087.098.098 0 0 0 .076 0s.087 0 .087-.109a.098.098 0 0 1-.087 0 .131.131 0 0 1 0-.12h-.141l-.055-.065a.054.054 0 0 1 0-.054.065.065 0 0 1 0-.065s0-.065.076-.065a.12.12 0 0 1 .087 0 .109.109 0 0 1-.108 0 .088.088 0 0 1 0-.076h.076a.097.097 0 0 1 .065.087s.087.076.109.14c.021.066 0-.075 0-.097a.087.087 0 0 0-.066-.087.175.175 0 0 1 .13.098.293.293 0 0 1 0 .141.548.548 0 0 0 0 .13c0 .11 0 .164.11.218a.466.466 0 0 0 .239.065h.13a.628.628 0 0 1 0-.206s.054 0 .087-.109a.336.336 0 0 1 0-.109.076.076 0 0 1-.065-.065l.087-.076c.087-.076 0-.076.087-.076a.195.195 0 0 1 0-.174.163.163 0 0 1 .12 0h.195a.307.307 0 0 1 .141 0 .36.36 0 0 1 .152.109v.087a.282.282 0 0 1 .13 0 .206.206 0 0 1-.108.141c0 .065 0 .163-.065.26-.066.099 0 0 0 .088.204-.04.411-.058.62-.054.38 0 .782 0 1.086.065h.51c.24 0 .761 0 .761-.153a.075.075 0 0 0 0-.065h-1.564c-.159 0-.316-.021-.468-.065a.25.25 0 0 1-.195-.25c0-.196.195-.228.5-.228h.184a.784.784 0 0 1 .315-.065h.598c.09.008.181-.011.26-.055 0 .087-.336.142-.64.142h.217c.14-.002.278-.02.413-.055a1.63 1.63 0 0 1 .402 0c.123.002.244.024.359.066h-.152a2.022 2.022 0 0 0-.566.12h-.065c-.243.076-.495.12-.75.13a1.196 1.196 0 0 1-.336 0 .196.196 0 0 1-.12-.087H39.4c-.12 0-.25 0-.25.12 0 .119.174.173.533.173h1.184c.522 0 .555.098.555.26 0 .164-.642.24-.924.262h.217c.127.01.254.01.38 0h.174a.13.13 0 0 0 .11-.055.173.173 0 0 1-.066.142.065.065 0 0 0 0 .065.076.076 0 0 1-.076 0 .064.064 0 0 1 0 .065.151.151 0 0 1-.098.098h-.065c-.011 0 0 .054 0 .13h-.142a.185.185 0 0 1 0-.13.118.118 0 0 0-.065.076s0 .065-.065.065 0 .065 0 .087a.11.11 0 0 1 0-.098s.076-.098.12-.12h-.076a.055.055 0 0 1-.055 0 .065.065 0 0 1 0-.064.173.173 0 0 0-.12.054.251.251 0 0 1 .153-.142h.228v-.076a.229.229 0 0 1 0-.12h-.12a.512.512 0 0 0-.173 0 .837.837 0 0 1-.478.11.663.663 0 0 1-.326-.087.396.396 0 0 1 0 .108v.066a.218.218 0 0 1 .086.13.262.262 0 0 1 0 .13h-.304v.066a.076.076 0 0 1 0 .087.087.087 0 0 1-.087-.066h-.108v.066h.086-.076a.11.11 0 0 1-.076 0 .087.087 0 0 1 0-.055.13.13 0 0 0 0 .076h-.065a.153.153 0 0 1 0-.065v-.054a.098.098 0 0 1 0-.076.173.173 0 0 1 .098 0 .227.227 0 0 0-.109 0 .12.12 0 0 1 0-.076.12.12 0 0 1 .109-.066c.065 0 .098 0 .13.098.033.098.055.087.098.087a.324.324 0 0 1 .141 0 .498.498 0 0 0-.173-.076l-.055-.054a.316.316 0 0 1 0-.261.36.36 0 0 1-.174.076h-.228a.305.305 0 0 0-.315.152.272.272 0 0 1-.26.13.143.143 0 0 1-.12.055v-.065c0 .087-.163.26-.294.26h-.076a.119.119 0 0 0 .065-.076.38.38 0 0 1-.152.066c-.098 0-.185 0-.195.195a.108.108 0 0 1 0-.098v.066a.131.131 0 0 1 0-.11h-.674a.24.24 0 0 1-.142.164c.076-.087 0-.141 0-.163l-.065.065a.098.098 0 0 0 0-.098v.076a.163.163 0 0 1-.098.066.12.12 0 0 1-.108-.11.098.098 0 0 1-.055.088.152.152 0 0 0 0 .13l-.108-.01Zm1.218-1.62s.065 0 0 .087 0 .109 0 .13a.087.087 0 0 0 .087-.065.065.065 0 0 0 0-.065v-.076h.054v-.054a.492.492 0 0 0 0-.055l-.163-.076h.055l-.033.174ZM38.313 39.808a.097.097 0 0 1-.087 0v.054a.086.086 0 0 1 0-.076h.098-.065a.108.108 0 0 0-.087 0 .086.086 0 0 1 0-.087.175.175 0 0 1 .098 0s.076 0 .109.087c.032.087.065.087.108.087a.262.262 0 0 1 .13 0 .445.445 0 0 1-.173-.25.52.52 0 0 1 .065-.26.523.523 0 0 0-.228.075.478.478 0 0 1-.228.077.728.728 0 0 1-.11.337c0 .086-.162.152-.216.152a.141.141 0 0 0 .065-.076h-.142c-.075 0-.13 0-.14.173a.23.23 0 0 1 0-.108v.054a.087.087 0 0 1 0-.087h-.185a.228.228 0 0 1-.142.141.12.12 0 0 0 0-.14h-.065c-.065 0 0-.077 0-.088a.12.12 0 0 1-.12.12.11.11 0 0 1-.097-.087.087.087 0 0 0 0 .065v.087a.151.151 0 0 0 0 .13.131.131 0 0 1-.087 0 .11.11 0 0 1 0-.097.098.098 0 0 1 0-.077v-.065a.228.228 0 0 1-.13 0s-.055 0-.066-.087c-.01-.087-.054 0-.076.055a.098.098 0 0 1 0-.076.12.12 0 0 1 .109-.066h.141c-.109 0-.13-.076-.141-.097h-.087.065a.12.12 0 0 1 .087 0h.065a.1.1 0 0 1 .065.065v.065a.131.131 0 0 0 .12 0 .934.934 0 0 0 .228-.076.174.174 0 0 1 .13 0h.196a.505.505 0 0 1 0-.109.217.217 0 0 1-.217-.184.347.347 0 0 1-.185-.272s-.065 0-.065-.087a.13.13 0 0 1-.109 0 .152.152 0 0 1-.13-.065.153.153 0 0 1 0-.153.12.12 0 0 0 0 .087.099.099 0 0 0 .065 0 .087.087 0 0 0 .087-.087.088.088 0 0 1-.087-.065.119.119 0 0 1 0-.108v.076a.108.108 0 0 1 0-.087.065.065 0 0 1 0-.065.066.066 0 0 1 .023-.05.065.065 0 0 1 .053-.016h.087a.142.142 0 0 1-.109 0 .076.076 0 0 1 0-.054v-.054h-.054a.077.077 0 0 1 .065 0 .076.076 0 0 1 .076.065c.054 0 .087.12.109.152v-.076a.22.22 0 0 0 0-.087.12.12 0 0 1 .12.087.207.207 0 0 1 0 .076.538.538 0 0 1 0 .13.335.335 0 0 0 0 .11c0 .108 0 .151.097.206.1.08.189.176.261.282l-.38-.826a1.369 1.369 0 0 1 0-.424v-.065s.076-.065.098-.141v-.13s.054 0 .086-.077a.208.208 0 0 1 .076-.065.196.196 0 0 1 0-.174h.38c.056.02.105.053.142.098v.087a.217.217 0 0 1 .12.065.175.175 0 0 1-.109.12c0 .065 0 .174-.076.282l-.065.12c.206-.053.418-.078.63-.076.233 0 .466.022.696.065.126.01.253.01.38 0 .467 0 .793-.12.793-.228 0-.109-.054 0-.184 0h-.859a1.862 1.862 0 0 1-.489 0c-.13 0-.206-.098-.206-.185s.195-.272.39-.315a.576.576 0 0 1 .305-.076h.63a.479.479 0 0 0 .24-.054c0 .087-.348.14-.642.14a.74.74 0 0 0 .218 0c.139-.001.278-.02.413-.053a1.63 1.63 0 0 1 .402 0c.119-.016.24-.016.358 0h-.152c-.196.008-.39.045-.576.108h-.065a2.86 2.86 0 0 1-.75.13 1.087 1.087 0 0 1-.326 0c-.076 0-.109-.054-.12-.097-.173.065-.25.13-.25.206 0 .076.11.109.37.109h.87c.358 0 .521.054.532.174a.196.196 0 0 1-.054.13c-.263.174-.578.25-.892.218h-.173c0 .076.065.108.14.152.058.03.11.07.153.12a.153.153 0 0 1 0 .064.163.163 0 0 1-.109.098s.076.076.12.076a.195.195 0 0 1-.098.066v.054a.087.087 0 0 1 0 .087.097.097 0 0 1-.076 0H39.4a.12.12 0 0 1-.076.12.087.087 0 0 0 0 .086.14.14 0 0 1-.076 0 .097.097 0 0 1 0-.076.108.108 0 0 1 .087-.076h-.141a.11.11 0 0 1 0-.076h.152-.12a.098.098 0 0 1 .098-.087h.076a.195.195 0 0 1 .098.087.109.109 0 0 0 .054 0l.087-.076a.63.63 0 0 0-.271-.13h-.196c-.119.016-.24.016-.359 0a.424.424 0 0 1 .163.413v.097a.228.228 0 0 1 .076.25h-.184v.066c0 .065 0 .065-.066.087a.098.098 0 0 1-.087 0 .066.066 0 0 0 .066 0l-.468-.044Zm-.717-1.891v.141a.077.077 0 0 0 0 .066.097.097 0 0 0 .087-.066.097.097 0 0 0 0-.076v-.076h.054l-.152-.087h-.054v.055l.065.043ZM10.764 7.205a.424.424 0 1 0-.424.424.435.435 0 0 0 .435-.424M10.612 6.075a.424.424 0 1 0-.848 0 .435.435 0 0 0 .424.424.424.424 0 0 0 .424-.424ZM10.851 5.564a.424.424 0 1 0-.413-.413.412.412 0 0 0 .413.413ZM11.981 5.499a.424.424 0 1 0 0-.847.424.424 0 0 0 0 .847ZM16.806 4.717a.435.435 0 0 0 .61-.609.435.435 0 0 0-.61 0 .467.467 0 0 0 0 .609ZM16.72 3.782a.424.424 0 0 0-.587-.609.424.424 0 0 0 .134.689.412.412 0 0 0 .453-.09M15.59 3.63a.435.435 0 0 0 0-.609.424.424 0 0 0-.723.304.424.424 0 0 0 .722.304ZM14.633 4.336a.424.424 0 0 0 0-.598.435.435 0 0 0-.598 0 .435.435 0 0 0 0 .609.435.435 0 0 0 .609 0M11.927 3.782l.326.75.435-.522h.054c.016.308.016.616 0 .924a.543.543 0 1 0 1.087 0 .566.566 0 0 0-.294-.478 4.906 4.906 0 0 1-.597-.565v-.066l.717.066-.294-.761-.532.456v-.793l-.728.283.489.586a.088.088 0 0 0 0 .066l-.663.054ZM16.84 5.075c.132.17.214.373.238.587a1.163 1.163 0 0 1-.445-.413l.097.989a.771.771 0 0 1 .348-.348c0 .641-.293.902-.521.957a.424.424 0 0 1-.5-.326c0-.294.195-.468.456-.207 0-.598-.435-.598-.652-.326a.88.88 0 0 0-.24-1.217l-.13-.065a.924.924 0 0 0 .218 1.25c-.402-.196-.76.14-.402.586.076-.38.402-.304.478 0 .076.305-.185.598-.598.718-.62.184-.815-.207-.956-.533.13-.098.445-.065.793.163l-.282-.935a.967.967 0 0 1-.576.587c.037-.286.117-.564.239-.826l-1.163.479c.262.122.504.283.717.478a.957.957 0 0 1-.826 0l.468.902c.097-.402.293-.641.445-.674.12.337.26.76-.304 1.087a.728.728 0 0 1-.935-.098c-.152-.26 0-.576.348-.348-.054-.554-.543-.554-.685-.13 0-.576-.141-.913-.75-1.087a.859.859 0 0 0 .479 1.087h.14c-.325 0-.684.293-.227.695 0-.37.293-.402.467-.174a.413.413 0 0 1-.109.576c-.184.12-.554.109-1.087-.326.168-.043.344-.043.511 0l-.63-.76c.056.195.056.402 0 .597a1.087 1.087 0 0 1-.24-.587l-.488.424a5.77 5.77 0 0 1 1.62 1.587 19.154 19.154 0 0 1 2.586-1.293 21.93 21.93 0 0 1 2.75-.891 5.61 5.61 0 0 1 0-2.25l-.653.065ZM6.797 50.284a.794.794 0 0 1-.804-.467 1.293 1.293 0 0 1 .152-1.174c.196-.184.316-.315.305-.456-.011-.141-.076-.185-.218-.283a.619.619 0 0 0-.337-.108.88.88 0 0 0-.815.739 2.108 2.108 0 0 1-.184-2.01 3.26 3.26 0 0 0 .13-1.979 3.01 3.01 0 0 1-1.206 1.826c.057-.366.02-.74-.109-1.087a5.292 5.292 0 0 1-.174-1.021 3.26 3.26 0 0 1 .717-2.065 1.24 1.24 0 0 0 .316-.63 1.184 1.184 0 0 0-.087-.337c-.087-.24-.174-.49.097-.728a.326.326 0 0 1 .163 0c.147.016.29.056.424.12.125.066.262.107.402.119 1.359 0 2.26-1.783 2.456-3.543a13.193 13.193 0 0 1 1.957-5.271l.087-.152c0-.315.086-.565.14-.804-.314.402-.608.858-.934 1.358a10.52 10.52 0 0 1-2.043 2.554 1.565 1.565 0 0 1-.467 1.185 3.415 3.415 0 0 1-2.391 1.086 2.097 2.097 0 0 1-1.348-.445h.098a1.814 1.814 0 0 0 1.706-1 2.073 2.073 0 0 1-.673-.141c-.237.129-.502.196-.772.195-.503 0-.985-.195-1.348-.543a1.153 1.153 0 0 1-.347-.859 1.087 1.087 0 0 0 .87.457c.156-.01.311-.046.455-.109a2.87 2.87 0 0 1-.543-1.739c0-2.173 1.772-4.01 3.489-5.76 1.51-1.695 3.01-3.14 3.01-4.749a1.445 1.445 0 0 0-1.293-1.619h-.174a2.239 2.239 0 0 0-1.304.413 1.403 1.403 0 0 1 .521 1.087 2.935 2.935 0 0 1-1.195 2.401c-.79.548-1.65.99-2.554 1.316-1.391.543-2.695 1.14-2.945 2.25a2.413 2.413 0 0 1 .522-1.892 4.967 4.967 0 0 1 1.423-1.087c.891-.51 1.815-1.032 1.967-2.271a3.705 3.705 0 0 1-2.5 1.706.957.957 0 0 1-.488-.12c.956 0 1.304-.945 1.673-1.934.37-.989.804-2.173 1.989-2.173.265.009.526.064.772.163a3.423 3.423 0 0 1 2-.826c.583.02 1.143.239 1.586.62a2.445 2.445 0 0 1 .858 1.934c0 2.054-1.706 3.749-3.347 5.433-1.64 1.685-3.184 3.196-3.184 5.065a1.663 1.663 0 0 0 1.445 1.869c.101.016.204.016.305 0 1.576 0 2.782-1.978 4.162-4.26a37.428 37.428 0 0 1 2.065-3.163 25.876 25.876 0 0 0 2.38-5.585 31.266 31.266 0 0 1 1.228-3.26c-.479.447-.85.996-1.087 1.608-.435.836-.837 1.64-1.673 1.717.434-.37.445-1.141.467-2.043a7.078 7.078 0 0 1 .239-2.087 2.554 2.554 0 0 0 .13-1.608 4.348 4.348 0 0 0-.728 2c-.217 1.086-.434 2.281-1.26 2.51.336-.413.217-.99.086-1.609a5.43 5.43 0 0 1-.152-1.032c-.002-.7.178-1.39.522-2a4.161 4.161 0 0 0 .5-1.76c-.337.13-.5.586-.674 1.086-.25.729-.543 1.565-1.521 1.565h-.076a2.174 2.174 0 0 0 .608-1.793 2.564 2.564 0 0 1 .74-2.043 2 2 0 0 1-1.37-1.652 3.043 3.043 0 0 1 1.728-.62c-.261-.412.108-.814.456-1.086l.066.087c.086.13.206.293.336.293.032.011.066.011.098 0 .8-.54 1.652-1 2.543-1.369.895-.33 1.81-.607 2.739-.826.217 0 .174-.282.141-.5V7.26h.272c.521 0 .793.173.837.543a2.793 2.793 0 0 1 1.608-.576 2.762 2.762 0 0 1 .098 2.271c.33.046.647.157.934.326.26.173.56.274.87.294h.26a1.315 1.315 0 0 1-1.086.598 3.696 3.696 0 0 1-.685-.087l-.424-.066a5.01 5.01 0 0 1 2.467 1.913c.457.565.848 1.087 1.413 1.087a1.26 1.26 0 0 1-.815.326 2.358 2.358 0 0 1-1.26-.544 2.468 2.468 0 0 0-1.24-.543 7.369 7.369 0 0 1 2.978 2.587c.305.434.5.717.826.728-.19.184-.442.292-.706.304a2.293 2.293 0 0 1-1.022-.38 4.559 4.559 0 0 0-1.087-.468c2.38 1.587 2.5 2.837 2.315 3.858a1.967 1.967 0 0 1 .468 1.554 2.24 2.24 0 0 0 0 .272 3.953 3.953 0 0 1 1.086-.391c.136-.029.27-.065.402-.109-.163.152-.315.326-.456.478a13.716 13.716 0 0 0-2.695 4.347c-.533-.532-1.087-1.086-1.315-1.271a1.164 1.164 0 0 0-.859-.424H20.6c.315.359.946 1.032 1.598 1.75.945 1.032 2.01 2.173 2.445 2.695.442.487.696 1.114.717 1.771 0 .12 0 .25.055.37.069.312.219.6.434.837.218.293.413.543.326.782a.652.652 0 0 0 .098.685c.087.125.14.271.152.423 0 .468-.326.772-1 .946a.957.957 0 0 0 .087-.891.87.87 0 0 0-.673-.413 1.489 1.489 0 0 1-.305 1.163 1.087 1.087 0 0 1-.793.347.924.924 0 0 1-.38-.086.565.565 0 0 0 .38-.305 1.358 1.358 0 0 0 0-1.087 1.36 1.36 0 0 1-1.185.87 1.087 1.087 0 0 1-.967-.717.695.695 0 0 0 .424.097.544.544 0 0 0 .543-.467c0-.12-.108-.217-.217-.326a.673.673 0 0 1-.293-.5.739.739 0 0 1 .554-.771c.26-.13.51-.25.51-.685 0-.435-.217-.5-.412-.5-.196 0-.565.087-.663.685-.315-.272-.261-.74-.218-1.25a2.543 2.543 0 0 0 0-1.087 1.152 1.152 0 0 1-.64 1.196c-.152.081-.285.192-.392.326-.196-.185-.196-.479-.196-.859a2.89 2.89 0 0 0-.217-1.195 1.64 1.64 0 0 1-.402.989 1.087 1.087 0 0 0-.25.521 3.457 3.457 0 0 0-1.51-.304 2.87 2.87 0 0 0-2.076.674c.434.326.87.63 1.304.913 1.413.945 2.64 1.76 2.64 3.423 0 .924-.804 1.543-1.51 2.087-.706.543-1.087.836-1.087 1.315a4.053 4.053 0 0 0 1.533 2.173 4.901 4.901 0 0 0 2.564 1.37.902.902 0 0 0 .576-.294.794.794 0 0 1 .468-.272h1.032a.653.653 0 0 1 .598.63 1.206 1.206 0 0 1 .902.772.978.978 0 0 1-.272 1.087.967.967 0 0 0-.848-.794c-.183.011-.361.063-.521.153a1.26 1.26 0 0 1 0 1.01 1.392 1.392 0 0 1-.924.696.977.977 0 0 0 .141-.913.37.37 0 0 0-.37-.185c-.238 0-.282.066-.347.13-.065.066-.174.218-.771.218h-.316c-.587 0-.608 0-.88-.413l-.087-.12c-.37-.521-.641-.76-.858-.76h-.13a.413.413 0 0 0-.197.348.565.565 0 0 0 .261.565c-1.315 0-1.51-.804-1.706-1.587-.13-.522-.26-1.022-.717-1.239a5.25 5.25 0 0 1 0 1.598 4.271 4.271 0 0 1-1.467-2.065 2.785 2.785 0 0 1-.076-.435c0-.37-.098-.685-.49-.88a.489.489 0 0 1-.314-.554c0-.153.097-.174.26-.229l.305-.12a1.805 1.805 0 0 0 1.086-1.814 3.107 3.107 0 0 0-1.64-2.934 8.335 8.335 0 0 1-.772 4.89 9.194 9.194 0 0 1-3.978 3.413l-.467.184c-.956.38-2.043.804-2.043 1.967 0 2.652 1.184 4.087 2.293 4.087a.782.782 0 0 0 .641-.37 1.087 1.087 0 0 1 .989-.467c.467 0 .989.174 1.087.521h.206c.5 0 1.022.305 1.022.804a.848.848 0 0 1-.696.957.543.543 0 0 0 0-.533.706.706 0 0 0-.576-.26.641.641 0 0 0-.25 0c-.152.097-.098.228 0 .413.07.142.106.298.109.456 0 .391-.109.826-.902 1.087a.534.534 0 0 0 .065-.479.413.413 0 0 0-.37-.163c-.217 0-.38.174-.543.36a.934.934 0 0 1-.684.423 3.414 3.414 0 0 1-.978-.163 2.24 2.24 0 0 0-.707-.12 1.903 1.903 0 0 0-.435 0c-.423.098-.521.272-.543.37a.304.304 0 0 0 .152.315l-.098.141Zm9.434-36.504a.468.468 0 0 1 .456.293.75.75 0 0 1-.24.99c-.173.162-.064.521.099 1.01a1.293 1.293 0 0 0 1.086.891c.14-.006.276-.04.402-.098a1.02 1.02 0 0 0 .61-.608 1.228 1.228 0 0 0-.077-.956c-.228-.479-.391-.794-.63-.794-.435 0-.707-.087-.848-.543-.141-.457.207-.62.609-.848l.217-.12c-.348-.108-.543-.238-.565-.369-.022-.13 0-.13.108-.196a.37.37 0 0 1 .229-.086c.108.009.214.035.315.076a.708.708 0 0 0 .217 0 .12.12 0 0 0 .109-.066c.076-.163-.109-.315-.272-.434l-.12-.098a.435.435 0 0 0-.293-.13c-.131.01-.26.043-.38.097l-2.174.848c-.304.12-.456.228-.467.49.005.057.005.115 0 .173 0 .206 0 .435.12.489.076 0 .13-.109.195-.228.066-.12.185-.38.402-.413h.066a.196.196 0 0 1 .173.087c.066.108 0 .348-.163.652h.229c.183-.062.372-.102.565-.12l.021.011Zm-1.511-2.978c-.8.326-1.62.598-2.456.816.372.08.717.255 1 .51a.859.859 0 0 0 .554.261h.076a.913.913 0 0 0 .663-.38 1.575 1.575 0 0 0 .217-1.228l-.054.021Zm4.119-1.782c-.64.368-1.31.684-2 .946l-.359.152a1.468 1.468 0 0 0 1.185.76.88.88 0 0 0 .554-.195.76.76 0 0 0 .283-.62c.021-.37.137-.73.337-1.043ZM69.547 44.416a1.837 1.837 0 1 1 0-3.674 1.837 1.837 0 0 1 0 3.674Zm0-3.195a1.369 1.369 0 1 0 0 2.738 1.369 1.369 0 0 0 0-2.738ZM61.94 46.394a1.66 1.66 0 0 1-.326 0 1.543 1.543 0 0 1-.902-.49 1.087 1.087 0 0 1-.294-.912c.057-.28.22-.525.457-.685.248-.173.545-.26.847-.25h.326c.406.072.768.298 1.011.63a.979.979 0 0 1 .174.772 1.085 1.085 0 0 1-.576.772 1.49 1.49 0 0 1-.717.163Zm-.207-1.902a1.021 1.021 0 0 0-.587.163.589.589 0 0 0-.26.38.556.556 0 0 0 .097.435 1.085 1.085 0 0 0 .728.435h.218a.794.794 0 0 0 .847-.544.163.163 0 0 0 0-.076.935.935 0 0 0-.826-.782l-.217-.011ZM51.213 20.583a.99.99 0 0 1-.804-.37.944.944 0 0 1-.185-.738c.06-.365.258-.692.555-.913.255-.2.567-.314.89-.326h.153c.255.032.488.16.652.358a.934.934 0 0 1 .185.729 1.403 1.403 0 0 1-.555.934c-.256.197-.567.311-.89.326Zm.468-1.88a1.022 1.022 0 0 0-.609.228.956.956 0 0 0-.391.696c-.001.11.033.216.098.304a.522.522 0 0 0 .434.185 1.01 1.01 0 0 0 .598-.217.978.978 0 0 0 .391-.62.618.618 0 0 0-.098-.391.51.51 0 0 0-.337-.174l-.086-.01ZM59.462 43.775a.978.978 0 0 1-.728-.337 1.477 1.477 0 0 1-.413-1.087c.004-.406.163-.795.445-1.087a1.02 1.02 0 0 1 1.226-.19c.147.077.274.187.372.32.207.296.314.65.304 1.011a1.576 1.576 0 0 1-.446 1.087 1.01 1.01 0 0 1-.76.283Zm0-2.391a.63.63 0 0 0-.402.185c-.2.206-.31.484-.305.771-.01.284.091.561.283.772a.608.608 0 0 0 .402.206.564.564 0 0 0 .424-.184c.2-.207.309-.485.304-.772a1.088 1.088 0 0 0-.282-.793.609.609 0 0 0-.403-.185h-.021ZM50.648 28.777a1.716 1.716 0 0 1-.739-.174 1.554 1.554 0 0 1-.804-.858 1.01 1.01 0 0 1 .652-1.348c.146-.052.301-.078.457-.076.25-.003.496.057.717.174.373.17.665.477.815.859.105.261.105.553 0 .815a1.022 1.022 0 0 1-.609.532 1.336 1.336 0 0 1-.489.076Zm-.38-1.967a1.01 1.01 0 0 0-.315 0 .555.555 0 0 0-.348.294.63.63 0 0 0 0 .576c.116.208.295.375.51.478.167.078.35.119.533.12a.914.914 0 0 0 .294 0 .588.588 0 0 0 .347-.294.522.522 0 0 0 0-.456 1.023 1.023 0 0 0-.565-.598 1.086 1.086 0 0 0-.456-.12Z"
-			/>
-			<path
-				fill="#fff"
-				d="M61.864 50.426h-.24c.076-.24.153-.468.207-.718.12-.456.26-.923.424-1.369a2.282 2.282 0 0 0-.772 1.272c-.152.456-.348.673-.587.673a.62.62 0 0 1-.358-.14 5 5 0 0 0 .434-.696 6.128 6.128 0 0 1 1.304-1.772l.11-.076a2.175 2.175 0 0 0 1.162-1.847 7.726 7.726 0 0 0 0-2.065c0-.305-.054-.62-.065-.957 0-1.01-.326-1.152-.63-1.152h-.424c-1 0-4.587-2.347-4.587-4.727a5.73 5.73 0 0 1 0-.739h-.076c-1.695 0-2.423.598-2.423 1.978a1.706 1.706 0 0 0 .445 1.347c.164.13.327.294.305.457-.022.163-.141.228-.37.337-.62.282-5.488 2.304-5.694 2.39-.13.056-.271.086-.413.088-.24 0-.348-.13-.49-.305l-.097-.12a.77.77 0 0 0-.576-.347.596.596 0 0 0-.304.098.216.216 0 0 0-.098.141c0 .087.065.185.141.283a.706.706 0 0 1 .196.39c0 .718-.663 1.088-1.935 1.088h-.532l.543-.935c.109-.217.218-.424.174-.51-.109 0-.228.173-.38.412l-.076.109-.663.858h-.109c-.215 0-.43-.021-.641-.065l1.14-1.087c.262-.271.5-.489.718-.717a3.26 3.26 0 0 1 2.076-.858h.152c1.37-.207 4.988-.74 4.988-1.978a1.804 1.804 0 0 0-.75-1.402 2.37 2.37 0 0 1-.913-1.934 2.933 2.933 0 0 1 .913-2.37c.826-.75 2.174-1.086 4.239-1.086h.293a3.989 3.989 0 0 0-2.75-2.228 4.032 4.032 0 0 1-2.434-1.652 4.924 4.924 0 0 1-.793-2.348h-1.815a8.854 8.854 0 0 1-1.521-.228 3.865 3.865 0 0 0-.696-.108.436.436 0 0 0-.326.097.455.455 0 0 0-.076.652c.065.102.141.232.228.392.24.413.576.989.87 1.412.293 0 .739.098.804.457a.457.457 0 0 1-.163.445.36.36 0 0 0-.13.185.282.282 0 0 0 0 .228.271.271 0 0 0 .434 0 .24.24 0 0 0 0-.206.892.892 0 0 1 .5-.109c.228 0 .455.033.674.098a4.792 4.792 0 0 1 1.087 1.739 6.9 6.9 0 0 0-1.663-.533c-.12 0-.185 0-.185.055 0 .054.065.108.207.152.388.107.756.28 1.086.51v.653a11.629 11.629 0 0 0-1.934-1.087h-.065a1.923 1.923 0 0 1-1.37-1.228 38.733 38.733 0 0 0-1.815-3.445c-.293-.478-.195-.674 0-.848l.326-.434.066-.098c.163-.272.26-.348.413-.348h1.054c.367.028.736-.016 1.087-.13l.304-.11a34.45 34.45 0 0 1-2.239-2.77.847.847 0 0 0-.63-.413.772.772 0 0 0-.522.163.513.513 0 0 0-.098.065l-.163.12a5.393 5.393 0 0 0-.271-.327l-.511-.554.815-.63c.05-.044.097-.092.141-.141a.663.663 0 0 1 .37-.283h.641a.74.74 0 0 1 .63.26c.402.4.839.764 1.304 1.087l.11.087h.064l.076.066.13-.12a.923.923 0 0 1 .653-.26.946.946 0 0 1 .663.325c.23.285.347.645.326 1.011 0 .13 0 .272-.076.402.355.29.717.605 1.086.946a7.738 7.738 0 0 1 1.685-3.684l-.13-.087c.017-.076.042-.149.076-.218.14-.445.358-1.152.51-1.673l.13-.435c.037-.084.066-.172.088-.26l.097-.327 1.435.565a.742.742 0 0 1 .13-.13 9.079 9.079 0 0 1 1.75-1.163c1.25-.707 2.554-1.446 2.554-2.565a1.739 1.739 0 0 0-.576-1.315 1.325 1.325 0 0 1 .174 1.01 1.695 1.695 0 0 1-.554.903 2.065 2.065 0 0 1-1.413.554 1.75 1.75 0 0 1-1-.446.848.848 0 0 0-.348-.195 1.02 1.02 0 0 0-.521.25c.054.902-.326 1.39-.696 1.467.108-.29.108-.611 0-.902-.12.065-.25.207-.272.706a1.39 1.39 0 0 1-1.162 1.272c.152-.228.358-1.087.217-1.174a.077.077 0 0 0-.065 0 .52.52 0 0 0-.24.13.456.456 0 0 1-.238.142.424.424 0 0 1-.359-.391c0-.076-.087-.087-.174-.087h-.185c-.076 0-.163 0-.195-.076a.152.152 0 0 1 0-.098c.22-.281.518-.49.858-.598.207-.087.891-.358 1-.608a.098.098 0 0 0 0-.11.282.282 0 0 0-.282-.151 5.097 5.097 0 0 0-1.826.76 1.142 1.142 0 0 1-.457.196.131.131 0 0 1-.098 0 .49.49 0 0 1 0-.326.824.824 0 0 0 0-.326c-.293-.663-.108-.815.196-1.087l.152-.13a12.854 12.854 0 0 1 2.641-1.728.837.837 0 0 1 .533-.793c.217-.076.434-.153.63-.24l.76-.282-6.14-7.716c-.173-.185-.13-.304-.097-.348a.12.12 0 0 1 .076 0 .543.543 0 0 1 .315.196l7.162 7.042a2.75 2.75 0 0 0 .152-1.695 3.261 3.261 0 0 1 1.75 2.043h.086a4.14 4.14 0 0 1 2 .478l.359.206c.208.152.45.25.706.283a.51.51 0 0 0 .304-.098.457.457 0 0 1-.065.337 1.24 1.24 0 0 1-.847.467 3.108 3.108 0 0 1 2.173 1.728c.26.478.5.891 1.087.978a.75.75 0 0 1-.609.272 1.435 1.435 0 0 1-.478-.087c.62.91.883 2.017.739 3.108 0 .728-.087 1.304.478 1.804a.623.623 0 0 1-.195 0 1.891 1.891 0 0 1-1.37-1.032c.199.822.225 1.677.076 2.51-.12.412-.12.85 0 1.26a.512.512 0 0 1-.184 0c-.413 0-.75-.412-.924-1.086-.098.394-.229.78-.391 1.152l.586.478-.228.24v.065c-.945 1-1.413 1.51-1.63 1.76l-.163.174-.185-.109a3.847 3.847 0 0 0-1.086 2.174c0 .848 1.086 2.38 2.282 3.89a1.435 1.435 0 0 1-.087-.478 1.022 1.022 0 0 1 1.032-1.021 1.641 1.641 0 0 1 1.522 1.575 1.086 1.086 0 0 1-1.087 1.022 1.49 1.49 0 0 1-.815-.293v.054a20.277 20.277 0 0 1 1.978 2.89c.89 1.86 1.619 2.652 2.445 2.652a1.935 1.935 0 0 0 1.956-1.89v-.098c0-2.022-1.271-3.152-2.619-4.347-1.347-1.196-2.815-2.543-2.815-4.945a3.162 3.162 0 0 1 1.087-2.456 3.054 3.054 0 0 1 1.88-.815 2.172 2.172 0 0 1 1.695.674c.138-.017.276-.017.413 0 1.392 0 1.576.988 1.75 1.945.174.956.304 1.64 1.24 1.76a1.347 1.347 0 0 1-.588.12c-.858 0-1.662-.74-2.097-1.934a3.13 3.13 0 0 0 2.173 2.749c.891.489 1.663.902 1.37 2.032 0-.64-.935-.989-2-1.402-1.522-.587-3.412-1.315-3.5-3.26a2.066 2.066 0 0 1 .511-1.565 1.316 1.316 0 0 0-.847-.272 2.521 2.521 0 0 0-2.293 2.478c0 2 1.282 3.13 2.64 4.347 1.359 1.217 2.826 2.5 2.826 4.804a2.869 2.869 0 0 1-1.021 2.304h.13c.49-.02.955-.231 1.293-.587a1.87 1.87 0 0 1-1.804 1.51 1.346 1.346 0 0 1-1.021-.423h-.218c-.144.01-.29.01-.434 0 .434.869.945 1.315 1.521 1.315h.163c-.33.282-.75.436-1.184.434a2.01 2.01 0 0 1-1.533-.76 2.324 2.324 0 0 1-.543-1.902c-.42-.5-.761-1.06-1.01-1.663 0-.13-.131-.206-.207-.206a.13.13 0 0 0-.11.054.566.566 0 0 0-.119.434 5.217 5.217 0 0 1-.587 2.174 4.425 4.425 0 0 0-.467 1.565c.042.504.253.978.598 1.348.195.302.492.526.837.63a.99.99 0 0 0 .456-.185.63.63 0 0 1 .326-.141h.087c.24.108.218.435.163 1.087 0 .39 0 1.25.087 2.25v.76c.104-.365.34-.68.663-.88.26-.176.567-.27.88-.272.094-.011.189-.011.283 0 .277.054.523.214.684.446.106.17.162.365.163.565.005.08.005.16 0 .239-.1.396-.349.739-.695.956-.26.185-.572.284-.891.283h-.109a1.086 1.086 0 0 1-.837-.468l-.076-.152v1.152c0 .859-.184 1.326-.652 1.576a.76.76 0 0 0-.684-.424.49.49 0 0 0-.511.402.554.554 0 0 0 .217.402.751.751 0 0 1 .294.63c.043.707-1.196 1.479-2.13 1.479Zm4.977-6.52c-.225.002-.444.07-.63.195a1.087 1.087 0 0 0-.479.674.576.576 0 0 0 .066.445.642.642 0 0 0 .565.26c.225.003.445-.066.63-.195.244-.148.42-.386.49-.663a.545.545 0 0 0-.077-.445.565.565 0 0 0-.38-.25h-.196l.01-.022Zm-6.39-7.043a1.012 1.012 0 0 0-.783.456.86.86 0 0 0 .066 1.196c.106.07.23.108.358.108a1.087 1.087 0 0 0 .793-.456.88.88 0 0 0-.065-1.195.727.727 0 0 0-.37-.11Zm2.173-3.413a1.01 1.01 0 0 0-.793.457.859.859 0 0 0 0 1.195.64.64 0 0 0 .37.109 1.021 1.021 0 0 0 .782-.446.88.88 0 0 0 0-1.206.598.598 0 0 0-.359-.098v-.01Zm.424-4.347a.576.576 0 0 0-.435.153.598.598 0 0 0-.173.423c.013.285.129.555.326.761.192.214.462.343.75.359a.554.554 0 0 0 .597-.565 1.087 1.087 0 0 0-.326-.761 1.087 1.087 0 0 0-.739-.359v-.01ZM54.213 17.15v.076a72.424 72.424 0 0 1-.718 2.293l7.608 4.347 1.684-1.815-.695-.554c.02.322-.01.644-.087.957a.773.773 0 0 1-.326-.674l-.565.706c.25-.007.497.074.695.228a.86.86 0 0 1-.608.294c-.22-.022-.43-.1-.609-.229-.5-.336-.674-.586-.543-.815a.359.359 0 0 1 .304-.174.467.467 0 0 1 .424.435.641.641 0 0 0 .38-.37.456.456 0 0 0 0-.39.424.424 0 0 0-.402-.196 1.084 1.084 0 0 0-.26 0 1.337 1.337 0 0 0 .793-1.543 1.391 1.391 0 0 0-1.24 1.271.597.597 0 0 0-.445-.565.532.532 0 0 0-.359.141.64.64 0 0 0-.152.5c.054-.01.11-.01.163 0a.445.445 0 0 1 .392.229.369.369 0 0 1 0 .369.401.401 0 0 1-.348.185 1.444 1.444 0 0 1-.652-.24.87.87 0 0 1-.38-1.162.335.335 0 0 1 .054-.109 1.012 1.012 0 0 1 .26.75l.61-1a.933.933 0 0 1-.349.054 1.228 1.228 0 0 1-.347-.054c.205-.276.509-.462.847-.521l-1.195-.696c.13.32.13.68 0 1a1.086 1.086 0 0 1-.348-.609l-.554 1.087c.173-.11.371-.173.576-.185a.519.519 0 0 1 .195 0 .901.901 0 0 1-.76.5 1.086 1.086 0 0 1-.5-.152c-.511-.283-.685-.554-.544-.826a.37.37 0 0 1 .326-.195.466.466 0 0 1 .435.39.674.674 0 0 0 .348-.401.469.469 0 0 0-.065-.38.435.435 0 0 0-.468-.218.957.957 0 0 0-.347.076 1.413 1.413 0 0 0 .478-1.663 1.37 1.37 0 0 0-.946 1.467c-.065-.293-.25-.489-.467-.489a.479.479 0 0 0-.359.152.62.62 0 0 0-.13.511.707.707 0 0 1 .206 0 .425.425 0 0 1 .38.218.336.336 0 0 1 0 .337.347.347 0 0 1-.303.163c-.24-.022-.471-.1-.674-.229a.86.86 0 0 1-.565-1.087c.232.096.424.269.543.49l.337-.837a.837.837 0 0 1-.424.12.76.76 0 0 1-.337-.077c.232-.228.5-.415.794-.554l-.826-.326.065-.01ZM49.55 21.17a.544.544 0 0 0-.336.13.989.989 0 0 0-.283.652v.066c0 .413.25.771.544.782a.5.5 0 0 0 .347-.141 1.021 1.021 0 0 0 .087-1.315.51.51 0 0 0-.337-.174h-.022Zm6.282-10.313c-.62 0-.685.358-.75.739a1.51 1.51 0 0 1-.141.478 5.49 5.49 0 0 1 2.26-.62 3.324 3.324 0 0 1-.62-.315 1.554 1.554 0 0 0-.749-.293v.01ZM110.822 9.477a4.827 4.827 0 0 1 3.565 1.206 4.577 4.577 0 0 1 1.206 3.38 4.584 4.584 0 0 1-1.206 3.39 4.695 4.695 0 0 1-3.565 1.196h-4.727v6.346h-2.065V9.477h6.792Zm-.685 7.433a3.619 3.619 0 0 0 2.576-.717 2.717 2.717 0 0 0 .815-2.174 2.704 2.704 0 0 0-.815-2.173 3.836 3.836 0 0 0-2.576-.729h-4.042v5.695l4.042.098Zm9.683-7.433v15.518h-1.847V9.477h1.847Zm12.759 15.475a2.51 2.51 0 0 1-1.326.282c-.4.01-.789-.134-1.087-.402a1.794 1.794 0 0 1-.423-1.315 4.35 4.35 0 0 1-1.772 1.315 5.78 5.78 0 0 1-2.173.402 5.73 5.73 0 0 1-1.446-.173 3.26 3.26 0 0 1-1.184-.544 2.462 2.462 0 0 1-.783-.967 3.086 3.086 0 0 1-.304-1.445c-.02-.54.092-1.078.326-1.565.211-.389.504-.726.858-.99a4.061 4.061 0 0 1 1.218-.575c.456-.13.923-.24 1.402-.326.51-.098.988-.174 1.445-.228.411-.038.818-.11 1.217-.218.302-.067.582-.21.815-.413a.989.989 0 0 0 .315-.793 1.96 1.96 0 0 0-.228-.978 1.812 1.812 0 0 0-.587-.565 2.473 2.473 0 0 0-.815-.261 5.992 5.992 0 0 0-.88-.065 3.976 3.976 0 0 0-1.956.445 1.93 1.93 0 0 0-.848 1.685h-1.837a3.903 3.903 0 0 1 .424-1.76c.267-.467.64-.866 1.087-1.164.453-.302.958-.52 1.489-.64a7.957 7.957 0 0 1 1.793-.197c.505.004 1.01.04 1.511.11.475.068.934.219 1.358.445.406.223.747.55.989.945.259.488.383 1.035.359 1.587v5.77c-.01.318.019.636.087.946 0 .207.217.304.5.304.191 0 .382-.03.565-.087l-.109 1.435Zm-2.978-5.76a2.17 2.17 0 0 1-.913.38c-.369.077-.771.142-1.184.196l-1.25.174a4.92 4.92 0 0 0-1.087.315 2.175 2.175 0 0 0-.815.609c-.22.318-.331.7-.315 1.087a1.55 1.55 0 0 0 .185.771c.121.2.284.37.478.5.208.137.44.233.685.283.27.062.548.091.826.086a4.187 4.187 0 0 0 1.565-.25 3.266 3.266 0 0 0 1.086-.619c.258-.23.465-.511.609-.826.128-.26.195-.546.195-.837l-.065-1.869Zm6.434-5.434v1.783a3.901 3.901 0 0 1 3.651-2.043 5.012 5.012 0 0 1 1.848.304 3.05 3.05 0 0 1 1.195.848c.311.375.531.818.641 1.293.132.544.194 1.103.185 1.662v7.39h-1.847v-7.607a2.177 2.177 0 0 0-.609-1.652 2.254 2.254 0 0 0-1.673-.608 3.765 3.765 0 0 0-1.468.26 2.846 2.846 0 0 0-1.086.74c-.272.32-.482.689-.62 1.086a4.859 4.859 0 0 0-.206 1.402v6.347h-1.848V13.758h1.837Zm12.041 0v1.783a3.901 3.901 0 0 1 3.652-2.043 5.01 5.01 0 0 1 1.847.304c.462.181.872.472 1.196.848.311.375.53.818.641 1.293.132.544.194 1.103.185 1.662v7.39h-1.848v-7.607a2.177 2.177 0 0 0-.609-1.652 2.254 2.254 0 0 0-1.673-.608 3.756 3.756 0 0 0-1.467.26 2.843 2.843 0 0 0-1.033.74c-.271.32-.481.689-.619 1.086a4.86 4.86 0 0 0-.207 1.402v6.347h-1.847V13.758h1.782Zm10.498-2.021v-2.26h1.848v2.26h-1.848Zm1.848 2.021v11.238h-1.848V13.758h1.848Zm4.608 0v1.783a3.902 3.902 0 0 1 3.695-2.043 5.01 5.01 0 0 1 1.847.304 3.06 3.06 0 0 1 1.196.848c.31.375.53.818.641 1.293.132.544.194 1.103.184 1.662v7.39h-1.847v-7.607a2.177 2.177 0 0 0-.609-1.652 2.254 2.254 0 0 0-1.673-.608 3.756 3.756 0 0 0-1.467.26 2.843 2.843 0 0 0-1.033.74c-.271.32-.481.689-.619 1.086a4.86 4.86 0 0 0-.207 1.402v6.347h-1.847V13.758h1.739Zm18.779 14.41a5.122 5.122 0 0 1-3.956 1.37 7.42 7.42 0 0 1-1.598-.174 4.746 4.746 0 0 1-1.477-.565 3.51 3.51 0 0 1-1.087-1.021 2.972 2.972 0 0 1-.478-1.544h1.847c.008.31.119.608.315.848.197.237.441.43.717.565.291.151.602.257.924.315.315.064.635.096.957.098a3.26 3.26 0 0 0 1.543-.315 2.813 2.813 0 0 0 1.087-.87 3.42 3.42 0 0 0 .597-1.336c.132-.563.194-1.14.185-1.717v-.74a3.069 3.069 0 0 1-1.445 1.533 4.467 4.467 0 0 1-2.054.49 4.979 4.979 0 0 1-2.174-.457 4.577 4.577 0 0 1-1.565-1.228 5.422 5.422 0 0 1-.945-1.804 7.51 7.51 0 0 1-.315-2.174 8.57 8.57 0 0 1 .26-2.054 5.774 5.774 0 0 1 .87-1.913 4.964 4.964 0 0 1 1.608-1.412 5.033 5.033 0 0 1 2.456-.555 4.19 4.19 0 0 1 1.956.468 3.254 3.254 0 0 1 1.391 1.402v-1.62h1.598V24.04a5.87 5.87 0 0 1-1.217 4.13Zm-2.358-5.053a2.94 2.94 0 0 0 1.032-1c.27-.43.465-.905.576-1.401a6.397 6.397 0 0 0 .185-1.565c.002-.491-.057-.98-.174-1.457a3.823 3.823 0 0 0-.554-1.293 2.894 2.894 0 0 0-.989-.924 2.947 2.947 0 0 0-1.478-.347 3.262 3.262 0 0 0-1.522.337 3.195 3.195 0 0 0-1.087.902 4.2 4.2 0 0 0-.586 1.304 6.524 6.524 0 0 0-.185 1.543c.001.51.052 1.02.152 1.521.099.485.276.951.522 1.38.241.407.577.75.978 1 .46.271.988.403 1.521.38.562.032 1.12-.1 1.609-.38ZM106.182 35.559v15.519h-2.065v-15.52h2.065Zm4.944 4.282v1.782a3.88 3.88 0 0 1 3.641-2.043 4.957 4.957 0 0 1 1.847.304c.465.181.879.472 1.207.848.302.38.521.822.641 1.293.126.545.189 1.103.185 1.663v7.39h-1.859V43.47a2.242 2.242 0 0 0-.597-1.652 2.297 2.297 0 0 0-1.674-.608 3.781 3.781 0 0 0-1.478.26 2.719 2.719 0 0 0-1.022.74c-.279.315-.49.685-.619 1.087a4.507 4.507 0 0 0-.217 1.401v6.347h-1.848V39.841h1.793Zm11.726 8.726c.168.279.404.51.685.674.293.179.618.3.956.359.358.075.722.111 1.087.108.306 0 .611-.022.913-.065.321-.026.636-.099.935-.217.26-.106.486-.283.652-.511.182-.248.274-.55.261-.858a1.217 1.217 0 0 0-.544-1.087 4.776 4.776 0 0 0-1.358-.598 21.42 21.42 0 0 0-1.761-.413 8.69 8.69 0 0 1-1.771-.522 3.48 3.48 0 0 1-1.359-.945 2.413 2.413 0 0 1-.543-1.695 2.743 2.743 0 0 1 .369-1.457c.255-.397.592-.734.989-.989a4.693 4.693 0 0 1 1.348-.565c.487-.12.987-.182 1.489-.184a8.791 8.791 0 0 1 1.75.163c.508.098.995.29 1.434.565.415.274.755.648.989 1.086.258.534.402 1.115.424 1.707h-1.837a1.84 1.84 0 0 0-.891-1.478 2.945 2.945 0 0 0-.815-.316 4.254 4.254 0 0 0-.902-.097c-.284 0-.567.022-.848.065-.266.04-.526.116-.771.228-.215.102-.404.25-.554.435-.153.203-.23.453-.218.706a1.011 1.011 0 0 0 .337.794c.248.237.545.42.869.532.376.153.766.27 1.163.348l1.304.293c.468.109.924.228 1.359.37.417.134.812.328 1.174.576.34.234.622.542.825.902.225.412.334.878.316 1.347a3.005 3.005 0 0 1-1.522 2.782 4.597 4.597 0 0 1-1.554.576 9.047 9.047 0 0 1-1.695.163 8.02 8.02 0 0 1-1.793-.195 4.413 4.413 0 0 1-1.5-.641c-.45-.3-.823-.702-1.087-1.174a3.95 3.95 0 0 1-.424-1.771h1.837c.012.35.109.694.282 1Zm11.368-8.726v1.521a2.76 2.76 0 0 1 1.424-1.358 5.428 5.428 0 0 1 2.173-.424 5.136 5.136 0 0 1 2.271.478 4.8 4.8 0 0 1 1.609 1.293c.435.563.762 1.2.967 1.88.22.733.33 1.495.326 2.26a7.591 7.591 0 0 1-.326 2.261 5.195 5.195 0 0 1-.945 1.859c-.436.532-.985.96-1.609 1.25a5.258 5.258 0 0 1-2.249.456c-.314 0-.627-.029-.935-.087a4.735 4.735 0 0 1-1-.283 3.943 3.943 0 0 1-.956-.532 2.743 2.743 0 0 1-.739-.837v5.781h-1.848V39.841h1.837Zm6.629 3.988a3.78 3.78 0 0 0-.63-1.337 3.127 3.127 0 0 0-1.087-.934 3.26 3.26 0 0 0-1.543-.348 3.26 3.26 0 0 0-1.587.37 3.148 3.148 0 0 0-1.086.967 4.17 4.17 0 0 0-.609 1.358 6.74 6.74 0 0 0-.185 1.543c.002.54.071 1.076.207 1.598.113.488.32.949.608 1.358a3.26 3.26 0 0 0 1.087.946 3.41 3.41 0 0 0 1.63.358 3.164 3.164 0 0 0 1.598-.37c.44-.226.815-.563 1.086-.977.264-.43.455-.9.566-1.391a7.834 7.834 0 0 0 .173-1.609 5.562 5.562 0 0 0-.228-1.532Zm12.302 6.531c-.921.664-2.038 1-3.173.957a5.677 5.677 0 0 1-2.337-.435 4.595 4.595 0 0 1-1.673-1.217 5.061 5.061 0 0 1-1-1.87 8.502 8.502 0 0 1-.38-2.369 6.818 6.818 0 0 1 .38-2.347 5.916 5.916 0 0 1 1.087-1.847 5.138 5.138 0 0 1 1.673-1.218 5.284 5.284 0 0 1 2.12-.434 4.596 4.596 0 0 1 2.488.62c.653.406 1.2.96 1.598 1.618.399.64.679 1.347.826 2.087a8.31 8.31 0 0 1 .185 2.174h-8.401c-.022.473.037.947.174 1.401.118.432.329.833.619 1.174.294.353.666.632 1.087.815a3.74 3.74 0 0 0 1.576.305 3.15 3.15 0 0 0 1.923-.544c.534-.4.889-.993.989-1.652h1.837a4.498 4.498 0 0 1-1.598 2.782Zm-.445-7.216a3.496 3.496 0 0 0-.674-1.01 3.37 3.37 0 0 0-1-.674 3.259 3.259 0 0 0-1.271-.25 3.06 3.06 0 0 0-2.293.934 3.258 3.258 0 0 0-.642 1.011 3.61 3.61 0 0 0-.282 1.228h6.444a3.672 3.672 0 0 0-.282-1.239Zm10.867-1.325a2.78 2.78 0 0 0-1.858-.587 3.503 3.503 0 0 0-1.674.37 3.047 3.047 0 0 0-1.086.977 4.462 4.462 0 0 0-.576 1.402 8.081 8.081 0 0 0-.174 1.641c0 .513.062 1.024.185 1.522.113.468.308.913.576 1.315.262.388.616.706 1.032.923.46.247.978.367 1.5.348a3.01 3.01 0 0 0 2.173-.717c.55-.537.887-1.255.946-2.021h1.901a5.175 5.175 0 0 1-1.564 3.26 4.838 4.838 0 0 1-3.402 1.087 5.755 5.755 0 0 1-2.271-.413 4.339 4.339 0 0 1-1.641-1.174 5.005 5.005 0 0 1-1.011-1.815 7.726 7.726 0 0 1-.337-2.315 8.323 8.323 0 0 1 .326-2.358 5.61 5.61 0 0 1 .989-1.913 4.754 4.754 0 0 1 1.663-1.282 5.586 5.586 0 0 1 2.325-.468 6.753 6.753 0 0 1 1.794.229c.533.14 1.034.38 1.478.706.439.325.809.735 1.086 1.206.288.535.462 1.123.511 1.728h-1.978a2.79 2.79 0 0 0-.913-1.651Zm9.759-1.978v1.63h-2.238v6.977a2.168 2.168 0 0 0 0 .521.564.564 0 0 0 .206.305c.128.08.273.129.424.14.235.014.471.014.706 0h.848v1.631h-1.413a6.691 6.691 0 0 1-1.228-.098c-.3-.05-.584-.173-.826-.358a1.578 1.578 0 0 1-.467-.74 4.36 4.36 0 0 1-.152-1.26v-7.118h-1.913v-1.63h1.913v-3.37h1.847v3.37h2.293Zm1.783 3.314a5.493 5.493 0 0 1 1.087-1.869 5.099 5.099 0 0 1 1.717-1.25 6.29 6.29 0 0 1 4.705 0 4.96 4.96 0 0 1 1.706 1.25c.491.54.861 1.176 1.087 1.87.24.747.358 1.529.348 2.314a7.149 7.149 0 0 1-.348 2.304 5.02 5.02 0 0 1-1.087 1.858 4.662 4.662 0 0 1-1.706 1.24 6.432 6.432 0 0 1-4.705 0 4.798 4.798 0 0 1-1.717-1.24 5.35 5.35 0 0 1-1.087-1.858 7.47 7.47 0 0 1-.348-2.304 7.603 7.603 0 0 1 .348-2.315Zm1.88 4.13c.158.488.413.939.75 1.326a3.254 3.254 0 0 0 2.478 1.087 3.26 3.26 0 0 0 1.358-.283c.419-.187.79-.466 1.087-.815.344-.382.6-.834.75-1.326a5.453 5.453 0 0 0 .271-1.815 5.56 5.56 0 0 0-.271-1.815 3.813 3.813 0 0 0-.75-1.337c-.3-.348-.671-.63-1.087-.825a3.38 3.38 0 0 0-1.358-.283 3.26 3.26 0 0 0-2.478 1.087c-.33.395-.584.849-.75 1.337a5.849 5.849 0 0 0-.272 1.814 5.76 5.76 0 0 0 .272 1.815v.033Zm12.53-7.444v2.369a5.067 5.067 0 0 1 1.652-2.022 4.206 4.206 0 0 1 2.478-.608v1.956a5.435 5.435 0 0 0-1.891.304 3.073 3.073 0 0 0-1.261.891 3.876 3.876 0 0 0-.695 1.424 7.937 7.937 0 0 0-.218 1.924v4.999h-1.847V39.84h1.782Zm15.519 11.193c-.408.213-.866.31-1.326.283-.4.009-.788-.135-1.086-.402a1.738 1.738 0 0 1-.424-1.315 4.35 4.35 0 0 1-1.772 1.315c-.692.272-1.43.409-2.173.402a5.74 5.74 0 0 1-1.446-.174 3.26 3.26 0 0 1-1.184-.544 2.59 2.59 0 0 1-.794-.967 3.271 3.271 0 0 1-.293-1.445c-.02-.54.092-1.078.326-1.565.211-.388.504-.726.859-.989.37-.259.782-.453 1.217-.576.456-.13.924-.239 1.402-.326.51-.098.989-.174 1.445-.228.408-.037.812-.11 1.206-.218.306-.065.59-.207.826-.412a.988.988 0 0 0 .305-.794c.015-.34-.064-.679-.229-.978a1.62 1.62 0 0 0-.586-.565 2.339 2.339 0 0 0-.805-.26 5.992 5.992 0 0 0-.88-.066 3.944 3.944 0 0 0-1.956.446 1.902 1.902 0 0 0-.848 1.684h-1.847a4.037 4.037 0 0 1 .434-1.76c.264-.47.637-.868 1.087-1.163.453-.303.958-.52 1.489-.642a7.95 7.95 0 0 1 1.793-.195c.506.003 1.01.04 1.511.108.475.07.935.22 1.358.446.406.221.744.548.978.946.269.484.397 1.033.37 1.586v5.77a4.61 4.61 0 0 0 .076.946c0 .207.228.305.511.305.191-.001.382-.03.565-.087l-.109 1.434Zm-2.999-5.76a2.178 2.178 0 0 1-.913.38c-.37.088-.772.142-1.185.196l-1.25.174a4.911 4.911 0 0 0-1.086.315 2.085 2.085 0 0 0-.815.62c-.224.316-.335.7-.316 1.087-.007.269.057.535.185.771.121.203.284.377.478.511.211.129.443.22.685.272.271.062.548.091.826.087a4.226 4.226 0 0 0 1.565-.25 3.261 3.261 0 0 0 1.087-.62 2.5 2.5 0 0 0 .608-.826c.123-.262.19-.547.196-.837l-.065-1.88Zm9.52-5.433v1.63h-2.239v6.977a2.168 2.168 0 0 0 0 .521.57.57 0 0 0 .206.305.936.936 0 0 0 .489.152c.236.013.472.013.707 0h.848v1.63h-1.413a6.677 6.677 0 0 1-1.228-.098 1.857 1.857 0 0 1-.826-.358 1.572 1.572 0 0 1-.468-.74 4.36 4.36 0 0 1-.152-1.26v-7.13h-1.923v-1.63h1.912v-3.368h1.848v3.369h2.239Zm9.998 10.52a5.052 5.052 0 0 1-3.173.956 5.751 5.751 0 0 1-2.337-.435 4.646 4.646 0 0 1-1.684-1.217 5.248 5.248 0 0 1-1.011-1.87 9.42 9.42 0 0 1-.381-2.369c-.008-.799.125-1.593.392-2.347a5.712 5.712 0 0 1 1.086-1.847 5.142 5.142 0 0 1 1.674-1.218 5.28 5.28 0 0 1 2.119-.434 4.597 4.597 0 0 1 2.489.62c.66.403 1.215.958 1.619 1.618.4.64.68 1.347.826 2.087.151.714.213 1.444.185 2.174h-8.412c-.016.473.043.946.174 1.401.127.431.342.83.631 1.174.295.351.667.63 1.086.815a3.793 3.793 0 0 0 1.587.305 3.155 3.155 0 0 0 1.924-.544c.529-.403.884-.994.989-1.652h1.825a4.41 4.41 0 0 1-1.608 2.782Zm-.446-7.217a3.108 3.108 0 0 0-.674-1.01 3.254 3.254 0 0 0-.999-.674 3.198 3.198 0 0 0-1.272-.25 3.262 3.262 0 0 0-1.293.25 3.357 3.357 0 0 0-.989.684 3.517 3.517 0 0 0-.652 1.011c-.155.393-.25.807-.283 1.228h6.456a3.445 3.445 0 0 0-.294-1.239ZM85.827 0h3.26v59.772h-3.26V0Z"
-			/>
-		</svg>
-	</div>
-</header>
-<main>
-	<div class="container">
-		<!-- Update page heading below  -->
-		<h1>Video title</h1>
-
-		<div class="video-container">
-			<!-- Update embed video iframe src below -->
-			<iframe
-				src="${youtubeUrlExample}"
-				allowfullscreen
-			></iframe>
-		</div>
-	</div>
-</main>
-</body>
-</html>
-`;
