+++ apps/functions/applications-background-jobs/malware-detected/index.js
@@ -1,6 +1,7 @@
 import { sendDocumentStateAction } from './src/back-office-api-client.js';
-import { isScannedFileHtml, isUploadedHtmlValid } from './src/html-validation.js';
-import { getPublishedStatusFromScanResult, guidFromBlobURI } from './src/util.js';
+import { isScannedFileHtml, isUploadedHtmlValid } from '../common/html-validation.js';
+import { getPublishedStatusFromScanResult } from './src/util.js';
+import { guidFromBlobURI } from '../common/util.js';
 
 /**
  * @type {import('@azure/functions').AzureFunction}
