+++ apps/functions/applications-background-jobs/malware-detected/src/back-office-api-client.js
@@ -1,6 +1,6 @@
 import { HTTPError } from 'got';
-import config from './config.js';
 import { requestWithApiKey } from '../../common/backend-api-request.js';
+import config from '../../common/config.js';
 
 /**
  *
