--- apps/functions/applications-background-jobs/malware-detected/src/blob-client.js
@@ -1,4 +1,0 @@
-import { BlobStorageClient } from '@pins/blob-storage-client';
-import config from './config.js';
-
-export const blobClient = BlobStorageClient.fromUrl(config.BLOB_STORAGE_ACCOUNT_HOST);
