--- apps/functions/applications-background-jobs/malware-detected/src/config.js
@@ -1,24 +1,0 @@
-import { loadEnvironment } from '@pins/platform';
-import joi from 'joi';
-
-const schema = joi.object({
-	NODE_ENV: joi.string().valid('development', 'production', 'test'),
-	API_HOST: joi.string(),
-	BLOB_SOURCE_CONTAINER: joi.string(),
-	BLOB_STORAGE_ACCOUNT_HOST: joi.string()
-});
-
-const environment = loadEnvironment(process.env.NODE_ENV);
-
-const { value, error } = schema.validate({
-	NODE_ENV: environment.NODE_ENV,
-	API_HOST: environment.API_HOST,
-	BLOB_SOURCE_CONTAINER: environment.BLOB_SOURCE_CONTAINER,
-	BLOB_STORAGE_ACCOUNT_HOST: environment.BLOB_STORAGE_ACCOUNT_HOST
-});
-
-if (error) {
-	throw error;
-}
-
-export default value;
