+++ apps/functions/applications-background-jobs/malware-detected/src/util.js
@@ -1,13 +1,4 @@
 /**
- * @param {string} uri
- * @returns {string | null}
- * */
-export const guidFromBlobURI = (uri) => {
-	const match = uri.match(/^.+\/document-service-uploads\/application\/.+?\/(.+?)\//);
-	return match?.[1] ?? null;
-};
-
-/**
  * @param {string} scanResult
  * @returns {string}
  * @throws {Error}
@@ -29,6 +20,7 @@ export const getPublishedStatusFromScanResult = (scanResult) => {
  */
 export const streamToString = async (readableStream) => {
 	return new Promise((resolve, reject) => {
+		/** @type {string[]} */
 		const chunks = [];
 		readableStream.on('data', (chunk) => {
 			chunks.push(chunk.toString());
@@ -38,12 +30,4 @@ export const streamToString = async (readableStream) => {
 		});
 		readableStream.on('error', reject);
 	});
-};
-
-/**
- * @param {string} uri
- * @returns {string}
- */
-export const extractBlobNameFromUri = (uri) => {
-	return new URL(uri).pathname.split('/').slice(2).join('/');
 };
