+++ apps/functions/applications-background-jobs/unpublish-document/utils.js
@@ -1,4 +1,4 @@
-import config from './config.js';
+import config from '../common/config.js';
 
 /**
  *
